chore: more code removal
