--- packages/shared/src/components/Rank.tsx
@@ -1,264 +1,0 @@
-import React, {
-  CSSProperties,
-  forwardRef,
-  HTMLAttributes,
-  LegacyRef,
-  ReactElement,
-  ReactNode,
-  useState,
-} from 'react';
-import classNames from 'classnames';
-import { getRank } from '../lib/rank';
-
-export interface RankProps extends HTMLAttributes<SVGElement> {
-  rank: number;
-  colorByRank?: boolean;
-}
-
-const outlinePaths: ReactNode = [
-  <path
-    key="rank-outline-1"
-    d="M14.9349954,3.5 C15.944078,3.5 16.8806603,4.00680277 17.4339665,4.84009984 L17.5394253,5.01105908 L20.5418806,10.2629023 C21.1187537,11.2719588 21.1548082,12.496536 20.6500443,13.5325326 L20.5418806,13.7370977 L17.5394253,18.9889409 C17.0386038,19.8649692 16.1337878,20.4265237 15.1357532,20.4932961 L14.9349954,20.5 L9.072009,20.5 C8.16261734,20.5 7.30754772,20.0879147 6.70807123,19.339568 L6.58359747,19.1740888 L6.46757913,18.9889409 L3.46512385,13.7370977 C2.94069394,12.8197739 2.861276,11.719922 3.26028307,10.6955589 L3.35311959,10.477302 L3.46512385,10.2629023 L6.46757913,5.01105908 C6.91892262,4.22157681 7.70105446,3.68377744 8.64825494,3.53476059 L8.85369263,3.50882913 L9.072009,3.5 L14.9349954,3.5 Z M14.9349954,5 L9.072009,5 L8.92639763,5.00706609 C8.49882606,5.04869564 8.10938189,5.27209782 7.85757538,5.61953797 L7.76979407,5.75552954 L4.76733879,11.0073727 L4.68675969,11.1638908 C4.4425806,11.6942671 4.4425806,12.3057329 4.68675969,12.8361092 L4.76733879,12.9926273 L7.76979407,18.2444705 L7.84819735,18.367375 C8.09654737,18.7179071 8.48377893,18.9451229 8.91038177,18.9912877 L9.072009,19 L14.9349954,19 L15.0806068,18.9929339 C15.5081784,18.9513044 15.8976225,18.7279022 16.149429,18.380462 L16.2372104,18.2444705 L19.2396656,12.9926273 C19.561982,12.428837 19.5888417,11.7473047 19.3202447,11.1638908 L19.2396656,11.0073727 L16.2372104,5.75552954 C15.9943878,5.33078856 15.5616248,5.05389933 15.0806068,5.00706609 L14.9349954,5 Z M12.0035022,7.099 L14.52107,11.5036864 C14.6749029,11.7727681 14.694132,12.0954905 14.5787573,12.3779824 L14.52107,12.4963136 L12.0035022,16.9 L9.4859344,12.4963136 L9.42824709,12.3779824 C9.32935456,12.1358465 9.32935456,11.8641535 9.42824709,11.6220176 L9.4859344,11.5036864 L12.0035022,7.099 Z M12.003,10.122 L10.93,11.999 L12.003,13.877 L13.076,11.999 L12.003,10.122 Z"
-    fill="var(--theme-text-tertiary)"
-  />,
-  <path
-    key="rank-outline-2"
-    d="M10.1376351,3.5 C10.4858026,3.5 10.827939,3.59088995 11.1302041,3.76368012 C11.5415106,3.99880388 11.8376606,4.35784654 11.9985158,4.7669186 C12.120225,4.45481267 12.3211418,4.17231813 12.5917125,3.95155898 L12.7327581,3.84647275 L12.8654642,3.76368012 C13.1173519,3.61968831 13.3969278,3.53257166 13.6846088,3.50753254 L13.8580333,3.5 L16.427518,3.5 C17.4366343,3.5 18.373243,4.00683636 18.9265411,4.8401781 L19.0319979,5.01114642 L22.0342279,10.2630042 C22.6110282,11.2720119 22.6470782,12.4964984 22.1423779,13.532441 L22.0342279,13.7369958 L19.0319979,18.9888535 C18.5311891,19.8649279 17.6263497,20.4265189 16.6282824,20.4932956 L16.427518,20.5 L13.8580333,20.5 C13.0125509,20.5 12.2896029,19.9753678 11.9969713,19.2338854 C11.7254592,19.9259356 11.0775904,20.4290853 10.305006,20.4930961 L10.1376351,20.5 L7.56815,20.5 C6.5590337,20.5 5.62242504,19.9931636 5.06912685,19.1598219 L4.96367011,18.9888536 L1.96144009,13.7369958 C1.38463982,12.7279881 1.34858981,11.5035016 1.85329006,10.467559 L1.96144011,10.2630042 L4.96367009,5.01114645 C5.46447889,4.1350721 6.36931833,3.5734811 7.3673856,3.50670439 L7.56815,3.5 L10.1376351,3.5 Z M10.1376351,5 L7.56815,5 L7.42253384,5.00706655 C6.99494846,5.04869882 6.60549416,5.27211538 6.35368989,5.61957439 L6.26591,5.75557323 L3.26368,11.007431 L3.18311107,11.163941 C2.93896281,11.6942891 2.93896281,12.3057109 3.18311107,12.836059 L3.26368,12.992569 L6.26591,18.2444268 L6.34431194,18.3673379 C6.59265942,18.7178889 6.9799009,18.9451193 7.40651746,18.9912871 L7.56815,19 L10.1376351,19 L10.2275107,18.9919443 C10.4607599,18.9496084 10.6376351,18.7454599 10.6376351,18.5 C10.6376351,18.4419721 10.6275362,18.3846143 10.6080043,18.3304333 L10.571715,18.2518577 L7.28153572,12.4962845 L7.22385569,12.3779595 C7.12497565,12.135837 7.12497565,11.864163 7.22385569,11.6220405 L7.28153572,11.5037155 L10.571715,5.74814226 L10.6093253,5.66611793 C10.6883289,5.44260963 10.5988758,5.18773797 10.3857773,5.06592003 C10.3353998,5.03712167 10.2805921,5.01742333 10.223861,5.00749092 L10.1376351,5 Z M16.427518,5 L13.8580333,5 C13.7709914,5 13.6854573,5.02272249 13.609891,5.06592003 C13.3967926,5.18773797 13.3073394,5.44260963 13.386343,5.66611793 L13.4239533,5.74814226 L16.7141326,11.5037155 C16.867946,11.7727842 16.8871727,12.0954832 16.7718126,12.3779595 L16.7141326,12.4962845 L13.4239533,18.2518577 C13.3807558,18.327424 13.3580333,18.4129581 13.3580333,18.5 C13.3580333,18.7454599 13.5349084,18.9496084 13.7681576,18.9919443 L13.8580333,19 L16.427518,19 L16.5731342,18.9929334 C17.0007195,18.9513012 17.3901738,18.7278846 17.6419781,18.3804256 L17.729758,18.2444268 L20.731988,12.992569 C21.0542637,12.428806 21.08112,11.7473239 20.8125569,11.163941 L20.731988,11.007431 L17.729758,5.75557323 C17.4869416,5.33080991 17.0541677,5.05390285 16.5731342,5.00706655 L16.427518,5 Z M11.9967272,6.23251447 L11.9462997,6.35389038 L11.8739549,6.49256903 L8.725,11.999 L11.8739549,17.507431 C11.9221922,17.5918133 11.9640466,17.679303 11.9993227,17.7691637 L12.042185,17.6617295 L12.1217134,17.507431 L15.269,11.999 L12.1217134,6.49256903 C12.0732124,6.40772538 12.0316093,6.32082434 11.9967272,6.23251447 Z"
-    fill="var(--theme-text-tertiary)"
-  />,
-  <path
-    key="rank-outline-3"
-    d="M10.1376351,3.5 C10.4858026,3.5 10.827939,3.59088995 11.1302041,3.76368012 C11.696863,4.08761125 12.050259,4.65186063 12.1236375,5.26245352 L12.4097147,4.75965265 C12.825148,4.03264433 13.5744359,3.56542079 14.4019044,3.50635546 L14.5803225,3.5 L17.1384102,3.5 C18.1197426,3.5 18.936,4.20676994 19.1059246,5.13905189 L22.0342279,10.2630042 C22.6110282,11.2720119 22.6470782,12.4964984 22.1423779,13.532441 L22.0342279,13.7369958 L19.0319979,18.9888535 C18.5311891,19.8649279 17.6263497,20.4265189 16.6282824,20.4932956 L16.427518,20.5 L13.8580333,20.5 C12.8363409,20.5 11.993581,19.7338989 11.8728663,18.7448093 L11.5902853,19.2403473 C11.174852,19.9673557 10.4255641,20.4345792 9.5980956,20.4936445 L9.41967746,20.5 L6.86158984,20.5 C5.88587508,20.5 5.07335074,19.8012987 4.8970571,18.8769364 L1.96144009,13.7369958 C1.43707614,12.819716 1.35766867,11.7199433 1.75662518,10.695638 L1.84944994,10.4773932 L1.96144011,10.2630042 L4.96367009,5.01114645 C5.41500206,4.22162286 6.19715171,3.68378937 7.14438217,3.53476286 L7.34982647,3.50882971 L7.56815,3.5 L10.1376351,3.5 Z M18.5119999,7.125 L16.3619999,10.888 L16.7141326,11.5037155 C16.867946,11.7727842 16.8871727,12.0954832 16.7718126,12.3779595 L16.7141326,12.4962845 L13.4239533,18.2518577 C13.3807558,18.327424 13.3580333,18.4129581 13.3580333,18.5 C13.3580333,18.7454599 13.5349084,18.9496084 13.7681576,18.9919443 L13.8580333,19 L16.427518,19 L16.5731342,18.9929334 C17.0007195,18.9513012 17.3901738,18.7278846 17.6419781,18.3804256 L17.729758,18.2444268 L20.731988,12.992569 C21.0542637,12.428806 21.08112,11.7473239 20.8125569,11.163941 L20.731988,11.007431 L18.5119999,7.125 Z M14.5803225,5 C14.2613387,5 13.9640938,5.15192973 13.7769323,5.40452097 L13.7120794,5.50386106 L6.42746827,18.2519305 C6.3842975,18.3274794 6.36158984,18.4129865 6.36158984,18.5 C6.36158984,18.7454599 6.538465,18.9496084 6.77171421,18.9919443 L6.86158984,19 L9.41967746,19 L9.53810485,18.9929784 C9.8112525,18.9604447 10.0593014,18.8164964 10.2230677,18.595479 L10.2879206,18.4961389 L17.5725317,5.74806947 C17.6013122,5.69770357 17.620998,5.64291176 17.6309241,5.58619833 L17.6384102,5.5 C17.6384102,5.25454011 17.461535,5.05039163 17.2282858,5.00805567 L17.1384102,5 L14.5803225,5 Z M7.56815,5 L7.42253384,5.00706655 C6.99494846,5.04869882 6.60549416,5.27211538 6.35368989,5.61957439 L6.26591,5.75557323 L3.26368,11.007431 L3.18311107,11.163941 C2.93896281,11.6942891 2.93896281,12.3057109 3.18311107,12.836059 L3.26368,12.992569 L5.48399988,16.878 L7.63499988,13.115 L7.28153572,12.4962845 L7.22385569,12.3779595 C7.12497565,12.135837 7.12497565,11.864163 7.22385569,11.6220405 L7.28153572,11.5037155 L10.571715,5.74814226 L10.6093253,5.66611793 C10.6883289,5.44260963 10.5988758,5.18773797 10.3857773,5.06592003 C10.3353998,5.03712167 10.2805921,5.01742333 10.223861,5.00749092 L10.1376351,5 L7.56815,5 Z"
-    fill="var(--theme-text-tertiary)"
-  />,
-  <path
-    key="rank-outline-4"
-    d="M10.1376351,3.5 C10.4858026,3.5 10.827939,3.59088995 11.1302041,3.76368012 C11.696555,4.0874352 12.0498748,4.65124738 12.1235176,5.26145803 L12.4097147,4.75965265 C12.825148,4.03264433 13.5744359,3.56542079 14.4019044,3.50635546 L14.5803225,3.5 L17.1384102,3.5 C18.1197426,3.5 18.936,4.20676994 19.1059246,5.13905189 L22.0342279,10.2630042 C22.6110282,11.2720119 22.6470782,12.4964984 22.1423779,13.532441 L22.0342279,13.7369958 L19.0319979,18.9888535 C18.5311891,19.8649279 17.6263497,20.4265189 16.6282824,20.4932956 L16.427518,20.5 L13.8580333,20.5 C12.8363409,20.5 11.993581,19.7338989 11.8728663,18.7448093 L11.5902853,19.2403473 C11.174852,19.9673557 10.4255641,20.4345792 9.5980956,20.4936445 L9.41967746,20.5 L6.86158984,20.5 C5.88587508,20.5 5.07335074,19.8012987 4.8970571,18.8769364 L1.96144009,13.7369958 C1.43707614,12.819716 1.35766867,11.7199433 1.75662518,10.695638 L1.84944994,10.4773932 L1.96144011,10.2630042 L4.96367009,5.01114645 C5.41500206,4.22162286 6.19715171,3.68378937 7.14438217,3.53476286 L7.34982647,3.50882971 L7.56815,3.5 L10.1376351,3.5 Z M20.731988,11.007431 L18.512,7.126 L16.361,10.889 L16.7141326,11.5037155 C16.867946,11.7727842 16.8871727,12.0954832 16.7718126,12.3779595 L16.7141326,12.4962845 L13.4239533,18.2518577 C13.3807558,18.327424 13.3580333,18.4129581 13.3580333,18.5 C13.3580333,18.7454599 13.5349084,18.9496084 13.7681576,18.9919443 L13.8580333,19 L16.427518,19 L16.5731342,18.9929334 C17.0007195,18.9513012 17.3901738,18.7278846 17.6419781,18.3804256 L17.729758,18.2444268 L20.731988,12.992569 C21.0542637,12.428806 21.08112,11.7473239 20.8125569,11.163941 L20.731988,11.007431 Z M9.0669873,15.580127 L8.50699988,14.611 L6.42746827,18.2519305 C6.39868775,18.3022964 6.37900196,18.3570882 6.36907593,18.4138017 L6.36158984,18.5 C6.36158984,18.7454599 6.538465,18.9496084 6.77171421,18.9919443 L6.86158984,19 L9.41967746,19 L9.53810485,18.9929784 C9.8112525,18.9604447 10.0593014,18.8164964 10.2230677,18.595479 L10.2879206,18.4961389 L11.5249999,16.33 L10.3660254,16.330127 L10.2210207,16.3231195 C9.79519822,16.2818335 9.40703093,16.0602505 9.15494744,15.7151982 L9.0669873,15.580127 Z M10.1376351,5 L7.56815,5 L7.42253384,5.00706655 C6.99494846,5.04869882 6.60549416,5.27211538 6.35368989,5.61957439 L6.26591,5.75557323 L3.26368,11.007431 L3.18311107,11.163941 C2.93896281,11.6942891 2.93896281,12.3057109 3.18311107,12.836059 L3.26368,12.992569 L5.48399988,16.878 L7.634,13.114 L7.28153572,12.4962845 L7.22385569,12.3779595 C7.12497565,12.135837 7.12497565,11.864163 7.22385569,11.6220405 L7.28153572,11.5037155 L10.571715,5.74814226 L10.6093253,5.66611793 C10.6883289,5.44260963 10.5988758,5.18773797 10.3857773,5.06592003 C10.3353998,5.03712167 10.2805921,5.01742333 10.223861,5.00749092 L10.1376351,5 Z M10.6547005,9.16987298 L10.5668028,9.17764841 C10.4806115,9.19303306 10.4004033,9.23074682 10.3339504,9.28630237 L10.2723704,9.347639 L10.2216878,9.41987298 L8.87638837,11.75 L8.83917321,11.8300094 C8.80940108,11.9123456 8.80195804,12.0006648 8.81684411,12.0859925 L8.83917321,12.1699906 L8.87638837,12.25 L10.2216878,14.580127 L10.2723704,14.652361 C10.3287896,14.7193125 10.4015547,14.769918 10.4828937,14.7996901 L10.5668028,14.8223516 L10.6547005,14.830127 L13.3452995,14.830127 L13.4331972,14.8223516 C13.5193885,14.8069669 13.5995967,14.7692532 13.6660496,14.7136976 L13.7276296,14.652361 L13.7783122,14.580127 L15.1236116,12.25 L15.1608268,12.1699906 C15.1905989,12.0876544 15.198042,11.9993352 15.1831559,11.9140075 L15.1608268,11.8300094 L15.1236116,11.75 L13.7783122,9.41987298 L13.7276296,9.347639 C13.6712104,9.2806875 13.5984453,9.23008203 13.5171063,9.2003099 L13.4331972,9.17764841 L13.3452995,9.16987298 L10.6547005,9.16987298 Z M17.1384102,5 L14.5803225,5 C14.2613387,5 13.9640938,5.15192973 13.7769323,5.40452097 L13.7120794,5.50386106 L12.4739999,7.669 L13.6339746,7.66987298 L13.7789793,7.67688046 C14.2048018,7.7181665 14.5929691,7.93974953 14.8450526,8.28480176 L14.9330127,8.41987298 L15.4919999,9.388 L17.5725317,5.74806947 C17.6157025,5.67252062 17.6384102,5.58701347 17.6384102,5.5 C17.6384102,5.25454011 17.461535,5.05039163 17.2282858,5.00805567 L17.1384102,5 Z"
-    fill="var(--theme-text-tertiary)"
-  />,
-  <path
-    key="rank-outline-5"
-    d="M11.0103641,1.67302444 C11.5786777,1.46795844 12.1953576,1.44517333 12.7744193,1.6046691 L12.9896362,1.67302444 L13.1120405,1.72340064 L13.3382053,1.82837669 L13.4552449,1.88909001 L15.7842895,3.23033188 C16.0724797,3.39629379 16.3037473,3.62314717 16.4711914,3.88531137 C16.776117,3.71575475 17.1177016,3.59737037 17.484651,3.53759926 L17.7077742,3.50882971 L17.9260977,3.5 L20.4955828,3.5 C20.8437503,3.5 21.1858867,3.59088995 21.4881518,3.76368012 C22.2814742,4.2171837 22.6568015,5.14171105 22.4309586,6.0049955 L22.3815234,6.1660137 L22.3307708,6.29131692 L22.2931605,6.37334125 L22.2319026,6.49256903 L19.083,11.999 L22.2319026,17.507431 C22.3470961,17.7089411 22.4258894,17.9281717 22.4656191,18.1550962 L22.4880502,18.3265755 L22.4955828,18.5 C22.4955828,19.4169705 21.8749265,20.2039179 21.0086796,20.4332377 L20.8533395,20.4678304 L20.7193685,20.4859551 L20.6294929,20.4940107 L20.4955828,20.5 L17.9260977,20.5 C17.4106313,20.5 16.9126187,20.3676007 16.4687148,20.1166118 C16.3358049,20.3257221 16.161494,20.5122649 15.9499058,20.6632753 L15.7842895,20.7696682 L13.4552448,22.11091 L13.3382053,22.1716233 L13.1120405,22.2765994 L12.9896362,22.3269756 C12.4213226,22.5320416 11.8046426,22.5548267 11.2255809,22.3953309 L11.0103641,22.3269756 L10.8879598,22.2765994 L10.6617949,22.1716233 L10.5446428,22.1108451 L8.21259456,20.7676045 C7.92673037,20.6029491 7.69688305,20.3783787 7.5297885,20.1188641 C7.15407983,20.3299308 6.73210731,20.4591871 6.29137333,20.49181 L6.06957021,20.5 L3.50008548,20.5 C2.39551598,20.5 1.50008548,19.6045695 1.50008548,18.5 C1.50008548,18.2098604 1.5632035,17.9239091 1.68423716,17.6617295 L1.7637656,17.507431 L4.911,11.999 L1.7637656,6.49256903 C1.26916644,5.62735731 1.4918966,4.54819118 2.244502,3.94285659 L2.37481035,3.84647275 L2.50751645,3.76368012 C2.75940406,3.61968831 3.03898001,3.53257166 3.32666097,3.50753254 L3.50008548,3.5 L6.06957021,3.5 C6.5901647,3.5 7.09146172,3.63489146 7.53078759,3.87991469 C7.66366396,3.67327996 7.83734152,3.48822797 8.04785995,3.33816898 L8.21259456,3.23239547 L10.5446428,1.88915487 L10.6617949,1.82837669 L10.8879598,1.72340064 L11.0103641,1.67302444 Z M12.1079873,13.4391535 C12.0053432,13.3884417 11.8846363,13.4150726 11.8120353,13.4966388 L11.7747768,13.5519915 L11.6762801,13.7369958 L8.77723158,18.7855494 C8.74939589,18.8340239 8.7305811,18.8852105 8.7202629,18.9371656 C8.68545681,19.1124241 8.74733308,19.2964266 8.88576653,19.4145665 L8.96126985,19.4678025 L11.2933181,20.8110431 L11.5194829,20.9160192 C11.6746442,20.9720064 11.8373222,21 12.0000001,21 L12.1623997,20.9906688 C12.2343298,20.9823744 12.305765,20.9685504 12.3760454,20.9491968 L12.4805173,20.9160192 L12.7066822,20.8110431 L15.0357268,19.4698013 C15.2482565,19.3474105 15.3369565,19.0925559 15.2575416,18.869392 L15.2197941,18.7875116 L12.3193878,13.7369958 L12.2207604,13.5518602 C12.1965532,13.5029359 12.1569259,13.4633319 12.1079873,13.4391535 Z M13.6216277,11.007431 L13.5410588,11.163941 C13.2969105,11.6942891 13.2969105,12.3057109 13.5410588,12.836059 L13.6216277,12.992569 L16.6238577,18.2444268 L16.7022596,18.3673379 C16.9506071,18.7178889 17.3378486,18.9451193 17.7644652,18.9912871 L17.9260977,19 L20.4955828,19 L20.5854584,18.9919443 C20.8187076,18.9496084 20.9955828,18.7454599 20.9955828,18.5 C20.9955828,18.4419721 20.9854839,18.3846143 20.965952,18.3304333 L20.9296627,18.2518577 L17.6394834,12.4962845 L17.5818034,12.3779595 C17.4829233,12.135837 17.4829233,11.864163 17.5818034,11.6220405 L17.6394834,11.5037155 L20.9296627,5.74814226 L20.967273,5.66611793 C21.0462766,5.44260963 20.9568235,5.18773797 20.743725,5.06592003 C20.6933475,5.03712167 20.6385398,5.01742333 20.5818087,5.00749092 L20.4955828,5 L17.9260977,5 L17.7804815,5.00706655 C17.3528962,5.04869882 16.9634419,5.27211538 16.7116376,5.61957439 L16.6238577,5.75557323 L13.6216277,11.007431 Z M3.50008548,5 C3.41304359,5 3.32750951,5.02272249 3.25194322,5.06592003 C3.03884478,5.18773797 2.94939165,5.44260963 3.02839524,5.66611793 L3.06600551,5.74814226 L6.35618482,11.5037155 C6.50999823,11.7727842 6.5292249,12.0954832 6.41386485,12.3779595 L6.35618482,12.4962845 L3.06600551,18.2518577 C3.02280797,18.327424 3.00008548,18.4129581 3.00008548,18.5 C3.00008548,18.7454599 3.17696064,18.9496084 3.41020985,18.9919443 L3.50008548,19 L6.06957021,19 L6.21518636,18.9929334 C6.64277175,18.9513012 7.03222604,18.7278846 7.28403032,18.3804256 L7.37181021,18.2444268 L10.3740402,12.992569 C10.6963159,12.428806 10.7231722,11.7473239 10.4546091,11.163941 L10.3740402,11.007431 L7.37181021,5.75557323 C7.12899382,5.33080991 6.69621992,5.05390285 6.21518636,5.00706655 L6.06957021,5 L3.50008548,5 Z M12.4805173,3.08398084 C12.2219152,2.9906688 11.9424334,2.97511679 11.6768714,3.03732482 L11.5194829,3.08398084 L11.2933181,3.1889569 L8.96126984,4.5321975 C8.74876903,4.65459622 8.66008529,4.90942818 8.73948943,5.13257592 L8.77723157,5.21445059 L11.6762801,10.2630042 L11.77483,10.4480085 L11.8040285,10.4937312 C11.8265675,10.5217528 11.8550549,10.5447276 11.8876806,10.5608465 C11.9287382,10.5811312 11.9726859,10.5890411 12.0152889,10.5860102 L12.0776554,10.5734376 C12.1178548,10.5598245 12.1545922,10.5359877 12.1836326,10.5033612 L12.2208911,10.4480085 L12.3193878,10.2630042 L15.2197941,5.21248838 C15.3419297,4.99981202 15.2904147,4.73492575 15.1095344,4.58198393 L15.0357268,4.53019876 L12.7066822,3.1889569 L12.4805173,3.08398084 Z"
-    fill="var(--theme-text-tertiary)"
-  />,
-  <path
-    key="rank-outline-6"
-    d="M11.0103639,1.67302444 C11.5786776,1.46795844 12.1953575,1.44517333 12.7744192,1.6046691 L12.9896361,1.67302444 L13.1120404,1.72340064 L13.3382052,1.82837669 L13.4552447,1.88909001 L15.7814188,3.22868163 C16.0694373,3.39385211 16.3020969,3.621099 16.4707209,3.88487165 C16.7760384,3.71568146 17.1176884,3.59735815 17.4846487,3.53759722 L17.7077741,3.50882971 L17.9260976,3.5 L20.4955826,3.5 C20.8437502,3.5 21.1858865,3.59088995 21.4881517,3.76368012 C22.2814741,4.2171837 22.6568014,5.14171105 22.4309585,6.0049955 L22.3815233,6.1660137 L22.3307707,6.29131692 L22.2931604,6.37334125 L22.2318495,6.49266181 L19.083,12 L22.2319025,17.507431 C22.4046927,17.8096961 22.4955826,18.1518324 22.4955826,18.5 C22.4955826,19.4169705 21.8749264,20.2039179 21.0086795,20.4332377 L20.8533394,20.4678304 L20.7193684,20.4859551 L20.6294928,20.4940107 L20.4955826,20.5 L17.9260976,20.5 C17.4106312,20.5 16.9126186,20.3676007 16.4687147,20.1166118 C16.3991476,20.2264121 16.3179973,20.3298917 16.2258451,20.4254749 L16.079438,20.5627306 L15.9654406,20.6503066 L15.8909751,20.7020343 L15.7837696,20.7699675 L13.4552447,22.11091 L13.3382052,22.1716233 L13.1120404,22.2765994 L12.9896361,22.3269756 C12.4213224,22.5320416 11.8046425,22.5548267 11.2255808,22.3953309 L11.0103639,22.3269756 L10.8879596,22.2765994 L10.6617948,22.1716233 L10.5446426,22.1108451 L8.21487582,20.7689171 C7.92927355,20.6049121 7.6981521,20.3798507 7.52998567,20.1186737 C7.15475354,20.3297298 6.73258305,20.4591335 6.29153646,20.4917988 L6.06957008,20.5 L3.50008536,20.5 C2.39551586,20.5 1.50008536,19.6045695 1.50008536,18.5 C1.50008536,18.2098604 1.56320338,17.9239091 1.68422829,17.6617448 L1.76375038,17.5074574 L4.911,12 L1.76376548,6.49256903 C1.21558474,5.53362604 1.54857334,4.31186086 2.50751633,3.76368012 C2.75940394,3.61968831 3.03897988,3.53257166 3.32666085,3.50753254 L3.50008536,3.5 L6.06957008,3.5 C6.59055571,3.5 7.09221485,3.63509423 7.53177786,3.88046763 C7.60063273,3.77219479 7.68115829,3.66985815 7.77259623,3.57519419 L7.91784262,3.43916982 L8.03050038,3.35254238 L8.1048457,3.3007881 L8.21316652,3.23206595 L10.5446427,1.88915487 L10.6617948,1.82837669 L10.8879596,1.72340064 L11.0103639,1.67302444 Z M11.4998396,17.4654261 L10.029,16.617 L8.77844751,18.7854876 L8.73699663,18.876402 C8.67222934,19.0705122 8.73237718,19.285764 8.88892901,19.4172267 L8.96184182,19.468132 L11.2933179,20.8110431 L11.5194828,20.9160192 C11.6746441,20.9720064 11.837322,21 12,21 L12.1623996,20.9906688 C12.2343297,20.9823744 12.3057649,20.9685504 12.3760453,20.9491968 L12.4805172,20.9160192 L12.7066821,20.8110431 L15.035207,19.4701006 L15.1096724,19.4183728 C15.2679166,19.2842716 15.3268233,19.0644621 15.2566336,18.8693262 L15.2185617,18.787529 L13.9669999,16.619 L12.4981787,17.4654579 C12.4209326,17.509954 12.3392242,17.5433253 12.2552848,17.5655719 L12.1279813,17.5905983 L11.9990049,17.5989377 C11.8697498,17.5989335 11.7404955,17.573899 11.6187728,17.5238342 L11.4998396,17.4654261 Z M17.7839999,12.75 L16.9969999,12.75 L16.997,14.2959892 C16.997,14.6138732 16.8461074,14.9102487 16.5949339,15.0975538 L16.496142,15.1625094 L15.2659999,15.87 L16.6114313,18.2208783 L16.6238576,18.2444268 L16.7022595,18.3673379 C16.950607,18.7178889 17.3378485,18.9451193 17.764465,18.9912871 L17.9260976,19 L20.4955826,19 L20.5854583,18.9919443 C20.8187075,18.9496084 20.9955826,18.7454599 20.9955826,18.5 C20.9955826,18.4419721 20.9854838,18.3846143 20.9659519,18.3304333 L20.9296626,18.2518577 L17.7839999,12.75 Z M6.21099988,12.75 L3.06600539,18.2518577 C3.02280785,18.327424 3.00008536,18.4129581 3.00008536,18.5 C3.00008536,18.7454599 3.17696052,18.9496084 3.41020973,18.9919443 L3.50008536,19 L6.06957008,19 L6.21518624,18.9929334 C6.64277163,18.9513012 7.03222592,18.7278846 7.2840302,18.3804256 L7.37181008,18.2444268 L8.72899988,15.869 L7.50080283,15.1615201 C7.22537757,15.0028435 7.04390115,14.7241714 7.00697337,14.4130516 L7,14.2950317 L6.99999988,12.75 L6.21099988,12.75 Z M11.999,8.13000659 C11.8697697,8.13000659 11.7405357,8.15503672 11.6188296,8.20508546 L11.4999115,8.26347498 L8.99987779,9.70349439 L8.90108231,9.76844941 C8.68129733,9.93234049 8.5382918,10.1797438 8.50597455,10.4519961 L8.499,10.570026 L8.499,13.426974 L8.50597455,13.5450039 C8.5382918,13.8172562 8.68129733,14.0646595 8.90108231,14.2285506 L8.99987779,14.2935056 L11.4999115,15.733525 L11.6188296,15.7919145 C11.7405357,15.8419633 11.8697697,15.8669934 11.999,15.8669934 L12.1279607,15.8586563 C12.21356,15.8475377 12.2980444,15.8252978 12.3791838,15.7919367 L12.4981053,15.7335542 L15.002,14.291 L15.0938763,14.2305425 C15.3136822,14.0666535 15.4567037,13.8192392 15.4890246,13.5469717 L15.496,13.428935 L15.496,10.568065 L15.4890246,10.4500283 C15.4567037,10.1777608 15.3136822,9.93034652 15.0938763,9.7664575 L14.9950716,9.7015042 L12.493,8.26 L12.3791838,8.20506326 C12.2980444,8.17170217 12.21356,8.14946233 12.1279607,8.13834368 L11.999,8.13000659 Z M3.50008536,5 C3.41304347,5 3.32750939,5.02272249 3.2519431,5.06592003 C3.03884466,5.18773797 2.94939153,5.44260963 3.02839512,5.66611793 L3.06600539,5.74814226 L6.21099988,11.25 L6.99999988,11.25 L6.99999988,9.70296829 C6.99999988,9.38510454 7.15087359,9.08874501 7.40202103,8.90143741 L7.50080283,8.83647987 L8.72799988,8.128 L7.37181008,5.75557323 C7.1289937,5.33080991 6.6962198,5.05390285 6.21518624,5.00706655 L6.06957008,5 L3.50008536,5 Z M20.4955826,5 L17.9260976,5 L17.7804814,5.00706655 C17.352896,5.04869882 16.9634417,5.27211538 16.7116375,5.61957439 L16.605,5.787 L15.2669999,8.127 L16.496142,8.83549056 C16.7715949,8.9941598 16.9530934,9.2728448 16.9900258,9.5839835 L16.997,9.70201076 L16.9969999,11.25 L17.7839999,11.25 L20.9296626,5.74814226 L20.9672729,5.66611793 C21.0462765,5.44260963 20.9568233,5.18773797 20.7437249,5.06592003 C20.6933474,5.03712167 20.6385397,5.01742333 20.5818086,5.00749092 L20.4955826,5 Z M12.4805172,3.08398084 C12.2219151,2.9906688 11.9424333,2.97511679 11.6768712,3.03732482 L11.5194828,3.08398084 L11.2933179,3.1889569 L8.96184181,4.53186798 L8.88749649,4.58362226 C8.72923854,4.71770859 8.67030441,4.9375159 8.74042639,5.13268674 L8.77846674,5.21450128 L10.0279999,7.38 L11.4998396,6.53257392 C11.6157045,6.46582239 11.7416112,6.42410098 11.870029,6.40740993 L11.9990049,6.39906234 L12.1279813,6.40740172 C12.2135939,6.41852363 12.2980912,6.44077022 12.3792418,6.47414156 L12.4981787,6.53254213 L13.9679999,7.379 L15.2185424,5.21245988 L15.2570946,5.130373 C15.3268417,4.93502813 15.2676744,4.71528989 15.1091452,4.58160326 L15.035207,4.52989946 L12.7066821,3.1889569 L12.4805172,3.08398084 Z"
-    fill="var(--theme-text-tertiary)"
-  />,
-  <path
-    key="rank-outline-7"
-    d="M10.7892869,1.75718262 C11.4990897,1.44284831 12.3012112,1.41665378 13.0264004,1.67859904 L13.2221448,1.75718262 L13.2941961,1.79140999 L13.5067104,1.90175083 L14.8022631,2.64899439 C15.0427875,2.78772307 15.2395258,2.97642012 15.3851679,3.19560666 C15.6249155,3.0933947 15.8827905,3.02953361 16.1487962,3.00807766 L16.3496688,3 L16.8714317,3 C17.4443902,3 17.9610763,3.2409307 18.3257047,3.62700676 C18.7257176,3.27185982 19.2341424,3.05267247 19.7753338,3.0083487 L19.9796688,3 L20.5014317,3 C21.6060012,3 22.5014317,3.8954305 22.5014317,5 C22.5014317,5.34869336 22.4102885,5.69028003 22.2385031,5.99127997 L22.1459387,6.13824271 L20.5360296,8.46319292 C20.6095574,8.49997268 20.6813821,8.54238427 20.7509434,8.59052964 C21.1802597,8.88767179 21.4521882,9.35912439 21.4980817,9.8738888 L21.505,10.0294865 L21.5047481,13.3322956 C21.492434,14.0041065 21.1031128,14.5806218 20.5421707,14.8636602 C22.3207267,17.5893237 22.8843936,19.7615246 21.440817,21.2051012 C20.0851104,22.5608078 18.0698776,22.1670301 15.5292645,20.5875536 L15.4553716,20.7228423 L15.3669458,20.8557653 C15.2434295,21.0244784 15.089373,21.1716788 14.9393173,21.271355 L13.4366996,22.1361897 L13.2221448,22.2428174 C12.5123419,22.5571517 11.7102205,22.5833462 10.9295334,22.2950372 L10.7172356,22.20859 L10.5040877,22.0978836 L9.18270248,21.3348814 C8.86623067,21.1521424 8.6284283,20.8854815 8.48117461,20.5792426 C5.94831005,22.1519092 3.93035812,22.5690833 2.56637601,21.2051012 C1.1223966,19.7611218 1.69314436,17.5876718 3.47158267,14.8536573 C3.37279012,14.8126253 3.28110111,14.755502 3.19368322,14.688889 C2.79811103,14.3874602 2.55161856,13.9338507 2.51039614,13.4545613 L2.50433501,13.2940206 L2.50487141,10.0430486 C2.50498165,9.35708984 2.89972863,8.76343056 3.47437158,8.47660479 L1.85561562,6.13841996 C1.62405592,5.80394484 1.5,5.40680892 1.5,5 C1.5,4.08302954 2.12065624,3.29608214 2.98690313,3.06676225 L3.14224327,3.03216959 L3.27621425,3.01404493 L3.36608988,3.00598926 L3.5,3 L4.02176289,3 C4.61759715,3 5.18873589,3.2125322 5.67072731,3.63224095 C5.91516368,3.37120348 6.23055226,3.17474307 6.58935282,3.07426481 L6.77224327,3.03216959 L6.90621425,3.01404493 L6.99608988,3.00598926 L7.13,3 L7.65176289,3 C7.98598336,3 8.31243361,3.06687132 8.61973847,3.19966873 C8.70763745,3.06565495 8.8164461,2.94214977 8.94450992,2.83397223 L9.07893085,2.7310824 L9.18252378,2.66545786 L10.5747321,1.8638103 L10.7892869,1.75718262 Z M12.0027158,17.949 L11.4433657,18.4221972 L10.9438377,18.827226 L10.4433016,19.2183238 L10.1968212,19.4055041 L9.91217842,19.6161683 C9.88325202,19.6368946 9.85908368,19.6635575 9.84128924,19.6943744 C9.78375393,19.7940154 9.80260923,19.9168027 9.87968769,19.9949191 L9.93277621,20.0358859 L11.2541614,20.798888 L11.3966649,20.8712883 C11.5904287,20.9570961 11.7980723,21 12.0057158,21 L12.1612316,20.9919555 C12.2646197,20.9812295 12.3671402,20.9597776 12.4670583,20.9275997 L12.6147667,20.8712883 L12.7572702,20.798888 L14.0791207,20.0371481 C14.1099862,20.0211481 14.1361159,19.9977071 14.1566318,19.9696841 C14.2245997,19.8768455 14.2191263,19.7527395 14.1507998,19.6669448 L14.1023462,19.620622 L13.8872893,19.4624402 L13.3045035,19.0172039 L13.0522401,18.8182988 L12.4446393,18.3234365 L12.0027158,17.949 Z M9.03011719,17.0038252 L8.87810407,16.9393983 L8.73039578,16.8631613 L8.31030674,17.1691 L8.10509384,17.3132788 L7.90372717,17.4512051 L7.5132447,17.7078356 L7.32448497,17.8263076 L7.14028363,17.9380626 L6.78626825,18.1409571 C6.1411412,18.4951796 5.60343409,18.7003678 5.21539613,18.7289677 L5.10351707,18.7322225 C4.92467626,18.7287259 4.78430897,18.6803935 4.68769635,18.5837808 C4.33356182,18.2296463 4.62811735,17.2876255 5.40173572,16.0178134 L5.52784239,15.8146307 L5.78953839,15.4133635 C5.91199574,15.2310464 6.04254916,15.0435831 6.18079354,14.8515948 L6.3938777,14.5602963 L6.73447013,14.111641 L6.91392341,13.8824708 L7.09931049,13.6503343 L7.48737335,13.1779485 L7.671,12.961 L7.199,12.5 L7.17082824,12.5004028 L6.76178442,12.9767355 L6.56399125,13.2129576 L6.37083097,13.4477185 L5.9988503,13.9123617 L5.6467248,14.3696751 L5.31533687,14.8186687 L5.15769525,15.0397361 L5.00556892,15.2583524 L4.71830335,15.6877362 C3.38565548,17.7366046 2.8724492,19.389854 3.62703618,20.144441 C3.85097819,20.368383 4.15406987,20.4806642 4.52316248,20.4930036 L4.66464032,20.4930036 C5.62007991,20.4620717 6.97107688,19.8259097 8.51727562,18.7630557 L8.79211975,18.5711003 L9.2720049,18.222052 L9.51575877,18.038349 L10.0101187,17.6532717 L10.4985315,17.25717 L10.995126,16.8392713 L11.1600305,16.7120122 L11.2690829,16.6227559 L11.3222832,16.5715025 C11.3307079,16.5607518 11.3298239,16.556335 11.3196313,16.5582519 L11.2994265,16.5647678 L11.159437,16.6318347 L10.6615373,16.9020104 C10.609047,16.9296062 10.5558892,16.9546369 10.5022075,16.9771461 L10.3397341,17.0371522 L10.2153387,17.0726343 L10.0358907,17.1088371 L9.88626522,17.1260942 L9.78567781,17.1312732 L9.61978194,17.1282939 L9.51035365,17.1193652 L9.33001892,17.0906173 L9.18673729,17.0554691 L9.03011719,17.0038252 Z M15.4913848,17.0213228 L15.2792304,16.8652023 L15.1360824,16.940101 L14.9947677,17.0002899 C14.9204251,17.0286128 14.8437091,17.0527349 14.7648536,17.0723253 L14.6186667,17.1029642 C14.556736,17.1135861 14.4945517,17.1212493 14.4323153,17.1259854 L14.2456505,17.1314449 C13.9349524,17.1260155 13.6279807,17.0482091 13.3498943,16.9020104 L12.9328218,16.645928 L12.7341487,16.5317785 L12.6560481,16.4927799 C12.618474,16.4766745 12.6098855,16.4803563 12.6302828,16.5038253 L12.6826179,16.5538693 L13.0975718,16.9122787 L13.5086615,17.25717 L13.9970744,17.6532717 L14.4914343,18.038349 L14.7351881,18.222052 C16.6754665,19.6675655 18.3767527,20.5300221 19.4840305,20.4930036 L19.6193208,20.4836687 C19.9276594,20.4508492 20.1842076,20.3403903 20.3801568,20.144441 C21.1328852,19.3917126 20.624058,17.7447432 19.2987219,15.7028684 L19.1481248,15.4743936 L18.8494978,15.0397361 L18.6918561,14.8186687 L18.3604682,14.3696751 L17.9171745,13.7968494 L17.4432018,13.2129576 L17.2454086,12.9767355 L16.8363648,12.5004028 L16.811,12.5 L16.336,12.963 L16.5198197,13.1779485 L17.0013098,13.766761 L17.4461142,14.3376483 L17.7741981,14.7793883 C17.8793187,14.9242207 17.9801416,15.0665514 18.0764959,15.2061185 L18.2176546,15.4133635 L18.4793506,15.8146307 C19.3462729,17.1868369 19.6922699,18.2110076 19.3194967,18.5837808 C19.2295738,18.6737037 19.1017459,18.7218011 18.9402755,18.7308537 L18.8393902,18.7316222 C18.3554478,18.7137496 17.6200501,18.4007222 16.7237161,17.8515928 L16.4963322,17.7094258 L16.10623,17.4531489 L15.9050626,17.3154123 L15.4913848,17.0213228 Z M7.73572816,15.2706238 L7.59808708,15.4546563 C7.49575931,15.5924353 7.39752055,15.7277775 7.30496651,15.8582675 L7.06,16.21 L7.42751162,15.956387 L7.80738777,15.6790565 C7.7694005,15.5457114 7.74539576,15.4088209 7.73572816,15.2706238 Z M16.2757098,15.2770942 C16.2656678,15.4157854 16.2411994,15.5514698 16.2035802,15.6823438 L16.3605106,15.7987762 L16.945,16.209 L16.7748928,15.9615345 L16.5699965,15.6737587 L16.2757098,15.2770942 Z M9.48527041,15.5655532 L9.54228197,15.594166 L9.60676153,15.616036 C9.64166346,15.6242039 9.67486144,15.6291136 9.70838791,15.630653 L9.79454368,15.6271412 L9.88022527,15.6083354 L9.96352565,15.5736429 L11.7730453,14.6223222 C11.8895749,14.5610589 12.0242517,14.5488063 12.1480403,14.5855642 L12.2383864,14.6223222 L14.0643642,15.5819124 C14.0816408,15.5900724 14.0962595,15.5961328 14.1110987,15.6014791 C14.1920393,15.6306407 14.2795369,15.6385564 14.3650991,15.6238814 C14.6070264,15.5823876 14.778337,15.3735477 14.780634,15.1364986 L14.7733807,15.0465547 L14.428,13.039 C14.4055381,12.9018673 14.4355026,12.7699957 14.5087142,12.6636247 L14.5715916,12.5890593 L16.0355237,11.1620781 C16.1142655,11.0853238 16.1655091,10.9847526 16.1813214,10.8759341 C16.2166181,10.6330253 16.0709374,10.4055641 15.8462002,10.3301273 L15.758417,10.3092314 L13.7353126,10.0152571 C13.6804974,10.007292 13.6281829,9.99041987 13.5801184,9.96591174 L13.5114281,9.92358146 C13.4688902,9.89237074 13.4312808,9.85442779 13.4002499,9.81095137 L13.3588437,9.74173649 L12.4540839,7.90849288 C12.417585,7.83453815 12.3635651,7.77115423 12.2972368,7.72355384 L12.2146739,7.6755289 L12.1836002,7.66327629 C12.0574848,7.61437658 11.9220918,7.6206818 11.8053681,7.67165801 C11.7240237,7.70867757 11.6550244,7.76292518 11.6032288,7.83308193 L11.5573478,7.90849288 L10.652588,9.74173649 C10.5943232,9.85979388 10.4925692,9.94886755 10.3708164,9.99189081 L10.2761191,10.0152571 L8.25301467,10.3092314 C8.14419615,10.3250437 8.04362494,10.3762873 7.96687064,10.4550291 L7.93211538,10.4946002 L7.90038202,10.53969 C7.79384466,10.7104196 7.80031806,10.9309471 7.91844846,11.0953599 L7.97590798,11.1620781 L9.4398401,12.5890593 C9.53411454,12.6809541 9.58738494,12.8052532 9.59067881,12.9343419 L9.58363842,13.031625 L9.23805092,15.0465547 C9.2194625,15.1549335 9.23711984,15.2664175 9.28828937,15.3637478 C9.31809746,15.4204461 9.35725835,15.4690489 9.40287988,15.508657 L9.48527041,15.5655532 Z M19.7740998,9.78017666 C19.7098977,9.77531882 19.6450913,9.79537939 19.5946008,9.83764376 L19.5494347,9.88720886 L18.22,11.808 L18.6317978,12.2937128 L19.0488288,12.8029288 L19.4434214,13.3028452 L19.555786,13.4500407 C19.5900211,13.4965625 19.6391143,13.5288917 19.6940047,13.5427345 L19.7504919,13.5502562 L19.8079245,13.5446883 C19.8966396,13.5255565 19.9681508,13.4592399 19.9943171,13.3728952 L20.005,13.3048057 L20.005,10.0294865 C20.005,9.94743956 19.9647416,9.87061499 19.8972776,9.82392122 C19.8783559,9.81082493 19.8582618,9.80069843 19.8375479,9.79344118 L19.7740998,9.78017666 Z M4.25491156,9.7933298 C4.13656483,9.7933298 4.03741153,9.87553369 4.01148328,9.98596914 L4.00487139,10.0432896 L4.00432149,13.2942681 C4.00351567,13.3732254 4.04001974,13.447942 4.10282528,13.4958002 C4.1394322,13.523695 4.18126647,13.5399373 4.22385438,13.5451127 L4.28774226,13.544649 C4.35107697,13.5360315 4.41107447,13.5031724 4.45263608,13.4480483 L4.5637716,13.3028452 L4.95836422,12.8029288 L5.37539526,12.2937128 L5.783,11.812 L4.4604396,9.90105643 C4.41375427,9.83360165 4.33694612,9.7933298 4.25491156,9.7933298 Z M7.16029645,9.1780963 L7.30327429,9.08532879 L7.46598352,8.99786792 L7.61601731,8.9345522 L4.84395508,4.93079002 L4.76845832,4.83481792 C4.60694788,4.65348202 4.38467023,4.53701899 4.14364391,4.50744457 L4.02176289,4.5 L3.5,4.5 L3.41012437,4.50805567 C3.17687516,4.55039163 3,4.75454011 3,5 C3,5.07627667 3.01744536,5.15119308 3.05059036,5.21916066 L3.0889039,5.28460499 L6.44196948,10.1290797 L6.44591934,10.1174806 C6.50272275,9.96255096 6.57899886,9.81342847 6.67498241,9.67376064 L6.77754378,9.53737946 L6.8927411,9.40868188 C6.96002666,9.33965399 7.03188799,9.27590913 7.10766951,9.21778133 L7.16029645,9.1780963 Z M20.5014317,4.5 L19.9796688,4.5 C19.692542,4.5 19.4214022,4.62325937 19.2329733,4.83481792 L19.1574766,4.93079002 L16.3878339,8.93125141 L16.5486786,8.99948718 L16.7043147,9.08244138 C17.056014,9.29188735 17.3315567,9.59944648 17.5003326,9.96152103 L17.5660392,10.1199064 L20.9125278,5.28460499 C20.9704177,5.20098621 21.0014317,5.10170223 21.0014317,5 C21.0014317,4.75454011 20.8245565,4.55039163 20.5913073,4.50805567 L20.5014317,4.5 Z M7.13,4.5 L7.04012437,4.50805567 C6.80687516,4.55039163 6.63,4.75454011 6.63,5 C6.63,5.07627667 6.64744536,5.15119308 6.68059036,5.21916066 L6.7189039,5.28460499 L9.066,8.675 L9.53971583,8.60667033 L10.155,7.36 L8.47395508,4.93079002 L8.39845832,4.83481792 C8.23694788,4.65348202 8.01467023,4.53701899 7.77364391,4.50744457 L7.65176289,4.5 L7.13,4.5 Z M16.8714317,4.5 L16.3496688,4.5 C16.062542,4.5 15.7914022,4.62325937 15.6029733,4.83481792 L15.5274766,4.93079002 L13.851,7.351 L14.4707158,8.60667033 L14.935,8.674 L17.2825278,5.28460499 C17.3404177,5.20098621 17.3714317,5.10170223 17.3714317,5 C17.3714317,4.75454011 17.1945565,4.55039163 16.9613073,4.50805567 L16.8714317,4.5 Z M12.0057158,3 C11.8499832,3 11.6942505,3.02413344 11.5443734,3.07240031 L11.3966649,3.12871167 L11.2541614,3.20111198 L9.93277155,3.96435288 C9.83817204,4.02984777 9.80274382,4.14891538 9.83904474,4.25247938 L9.86953325,4.31220474 L11.235227,6.28478564 L11.2552756,6.27588211 L11.394339,6.2261811 L11.5578622,6.18117095 C11.8887007,6.10495497 12.2404339,6.11219491 12.5796281,6.21422853 L12.767,6.281 L14.1391495,4.29863374 L14.1699459,4.22779037 C14.192069,4.14283008 14.1676247,4.05175947 14.1059314,3.98929643 L14.0528229,3.94835555 L12.7572702,3.20111198 L12.6147667,3.12871167 C12.4694439,3.06435583 12.3163137,3.02413344 12.1612316,3.00804448 L12.0057158,3 Z"
-    fill="var(--theme-text-tertiary)"
-  />,
-];
-
-const rankPaths: ((fill: string) => ReactNode)[] = [
-  (fill) => (
-    <>
-      <path
-        d="M9.072009,19 C8.58275684,19 8.12759112,18.7617236 7.84819735,18.367375 L7.76979407,18.2444705 L4.76733879,12.9926273 C4.44502239,12.428837 4.41816269,11.7473047 4.68675969,11.1638908 L4.76733879,11.0073727 L7.76979407,5.75552954 C8.01261659,5.33078856 8.44537962,5.05389933 8.92639763,5.00706609 L9.072009,5 L14.9349954,5 C15.4731728,5 15.9701056,5.28831447 16.2372104,5.75552954 L19.2396656,11.0073727 C19.5912835,11.6224167 19.5912835,12.3775833 19.2396656,12.9926273 L16.2372104,18.2444705 C15.9701056,18.7116855 15.4731728,19 14.9349954,19 L9.072009,19 Z M12.0035022,7.099 L9.4859344,11.5036864 C9.33210157,11.7727681 9.31287247,12.0954905 9.42824709,12.3779824 L9.4859344,12.4963136 L12.0035022,16.9 L14.52107,12.4963136 C14.696879,12.1887917 14.696879,11.8112083 14.52107,11.5036864 L12.0035022,7.099 Z"
-        fill={fill}
-      />
-      <path
-        d="M11.9996254,5 L11.9996254,7.104 L9.4859344,11.5036864 C9.33210157,11.7727681 9.31287247,12.0954905 9.42824709,12.3779824 L9.4859344,12.4963136 L11.9996254,16.894 L11.9996254,19 L9.072009,19 C8.58275684,19 8.12759112,18.7617236 7.84819735,18.367375 L7.76979407,18.2444705 L4.76733879,12.9926273 C4.44502239,12.428837 4.41816269,11.7473047 4.68675969,11.1638908 L4.76733879,11.0073727 L7.76979407,5.75552954 C8.01261659,5.33078856 8.44537962,5.05389933 8.92639763,5.00706609 L9.072009,5 L11.9996254,5 Z"
-        fillOpacity="0.2"
-        fill="var(--theme-rank-highlight)"
-      />
-      <path
-        d="M14.9349954,5 C15.4731728,5 15.9701056,5.28831447 16.2372104,5.75552954 L19.2396656,11.0073727 C19.4156449,11.3151927 19.5035493,11.6581112 19.5033791,12.0009958 L14.6529262,12.0009958 C14.6530969,11.8293876 14.6091449,11.6577453 14.52107,11.5036864 L12.0035022,7.099 L9.4859344,11.5036864 C9.39820028,11.6571494 9.35424821,11.8280599 9.35407818,11.9990042 L4.50362537,12 C4.50362537,11.7145395 4.56467014,11.4290789 4.68675969,11.1638908 L4.76733879,11.0073727 L7.76979407,5.75552954 C8.01261659,5.33078856 8.44537962,5.05389933 8.92639763,5.00706609 L9.072009,5 L14.9349954,5 Z"
-        fillOpacity="0.2"
-        fill="var(--theme-rank-highlight)"
-      />
-      <path
-        d="M11.9986254,5 L6.26856643,8.38145113 L7.76979407,5.75552954 C8.01261659,5.33078856 8.44537962,5.05389933 8.92639763,5.00706609 L9.072009,5 L11.9986254,5 Z"
-        fillOpacity="0.64"
-        fill="#FFFFFF"
-      />
-    </>
-  ),
-  (fill) => (
-    <>
-      <path
-        d="M10.1376351,5 L7.56815,5 C7.02995464,5 6.53300802,5.28833357 6.26591,5.75557323 L3.26368,11.007431 C2.9121065,11.6224452 2.9121065,12.3775548 3.26368,12.992569 L6.26591,18.2444268 C6.53300802,18.7116664 7.02995464,19 7.56815,19 L10.1376351,19 C10.4137774,19 10.6376351,18.7761424 10.6376351,18.5 C10.6376351,18.4129581 10.6149126,18.327424 10.571715,18.2518577 L7.28153572,12.4962845 C7.10574897,12.1887774 7.10574897,11.8112226 7.28153572,11.5037155 L10.571715,5.74814226 C10.7087602,5.50840651 10.6255131,5.20296522 10.3857773,5.06592003 C10.310211,5.02272249 10.2246769,5 10.1376351,5 Z M13.8580333,5 L16.427518,5 C16.9657134,5 17.46266,5.28833357 17.729758,5.75557323 L20.731988,11.007431 C21.0835615,11.6224452 21.0835615,12.3775548 20.731988,12.992569 L17.729758,18.2444268 C17.46266,18.7116664 16.9657134,19 16.427518,19 L13.8580333,19 C13.5818909,19 13.3580333,18.7761424 13.3580333,18.5 C13.3580333,18.4129581 13.3807558,18.327424 13.4239533,18.2518577 L16.7141326,12.4962845 C16.8899194,12.1887774 16.8899194,11.8112226 16.7141326,11.5037155 L13.4239533,5.74814226 C13.2869081,5.50840651 13.3701553,5.20296522 13.609891,5.06592003 C13.6854573,5.02272249 13.7709914,5 13.8580333,5 Z"
-        fill={fill}
-      />
-      <path
-        d="M10.1376351,5 C10.2246769,5 10.310211,5.02272249 10.3857773,5.06592003 C10.5988758,5.18773797 10.6883289,5.44260963 10.6093253,5.66611793 L10.571715,5.74814226 L7.28153572,11.5037155 C7.12772231,11.7727842 7.10849564,12.0954832 7.22385569,12.3779595 L7.28153572,12.4962845 L10.571715,18.2518577 C10.6149126,18.327424 10.6376351,18.4129581 10.6376351,18.5 C10.6376351,18.7454599 10.4607599,18.9496084 10.2275107,18.9919443 L10.1376351,19 L7.56815,19 C7.07888149,19 6.62370285,18.7617078 6.34431194,18.3673379 L6.26591,18.2444268 L3.26368,12.992569 C2.9414043,12.428806 2.91454799,11.7473239 3.18311107,11.163941 L3.26368,11.007431 L6.26591,5.75557323 C6.50872639,5.33080991 6.94150028,5.05390285 7.42253384,5.00706655 L7.56815,5 L10.1376351,5 Z"
-        fillOpacity="0.2"
-        fill="var(--theme-rank-highlight)"
-      />
-      <path
-        d="M10.1376351,5 C10.2246769,5 10.310211,5.02272249 10.3857773,5.06592003 C10.5988758,5.18773797 10.6883289,5.44260963 10.6093253,5.66611793 L10.571715,5.74814226 L7.28153572,11.5037155 C7.18904327,11.6655143 7.14521691,11.8467056 7.15005663,12.0268841 L3.00018036,12.0268841 C2.99622748,11.7325006 3.05720438,11.4374403 3.18311107,11.163941 L3.26368,11.007431 L6.26591,5.75557323 C6.50872639,5.33080991 6.94150028,5.05390285 7.42253384,5.00706655 L7.56815,5 L10.1376351,5 Z M13.8580329,5 C13.7709911,5 13.685457,5.02272249 13.6098907,5.06592003 C13.3967922,5.18773797 13.3073391,5.44260963 13.3863427,5.66611793 L13.423953,5.74814226 L16.7141323,11.5037155 C16.8066247,11.6655143 16.8504511,11.8467056 16.8456114,12.0268841 L20.9954876,12.0268841 C20.9994405,11.7325006 20.9384636,11.4374403 20.8125569,11.163941 L20.731988,11.007431 L17.729758,5.75557323 C17.4869416,5.33080991 17.0541677,5.05390285 16.5731342,5.00706655 L16.427518,5 L13.8580329,5 Z"
-        fillOpacity="0.2"
-        fill="var(--theme-rank-highlight)"
-      />
-      <path
-        d="M10.1376351,5 C10.2246769,5 10.310211,5.02272249 10.3857773,5.06592003 C10.5988758,5.18773797 10.6883289,5.44260963 10.6093253,5.66611793 L10.571715,5.74814226 L4.764795,8.3815021 L6.26591,5.75557323 C6.50872639,5.33080991 6.94150028,5.05390285 7.42253384,5.00706655 L7.56815,5 L10.1376351,5 Z"
-        fillOpacity="0.64"
-        fill="#FFFFFF"
-      />
-    </>
-  ),
-  (fill) => (
-    <>
-      <path
-        d="M18.5119999,7.125 L20.731988,11.007431 C21.0835615,11.6224452 21.0835615,12.3775548 20.731988,12.992569 L17.729758,18.2444268 C17.46266,18.7116664 16.9657134,19 16.427518,19 L13.8580333,19 C13.5818909,19 13.3580333,18.7761424 13.3580333,18.5 C13.3580333,18.4129581 13.3807558,18.327424 13.4239533,18.2518577 L16.7141326,12.4962845 C16.8899194,12.1887774 16.8899194,11.8112226 16.7141326,11.5037155 L16.3619999,10.888 L18.5119999,7.125 Z M17.1384102,5 C17.4145525,5 17.6384102,5.22385763 17.6384102,5.5 C17.6384102,5.58701347 17.6157025,5.67252062 17.5725317,5.74806947 L10.2879206,18.4961389 C10.1098778,18.8077139 9.77853431,19 9.41967746,19 L6.86158984,19 C6.58544746,19 6.36158984,18.7761424 6.36158984,18.5 C6.36158984,18.4129865 6.3842975,18.3274794 6.42746827,18.2519305 L13.7120794,5.50386106 C13.8901222,5.19228606 14.2214657,5 14.5803225,5 L17.1384102,5 Z M10.1376351,5 C10.2246769,5 10.310211,5.02272249 10.3857773,5.06592003 C10.5988758,5.18773797 10.6883289,5.44260963 10.6093253,5.66611793 L10.571715,5.74814226 L7.28153572,11.5037155 C7.12772231,11.7727842 7.10849564,12.0954832 7.22385569,12.3779595 L7.28153572,12.4962845 L7.63499988,13.115 L5.48399988,16.878 L3.26368,12.992569 C2.9414043,12.428806 2.91454799,11.7473239 3.18311107,11.163941 L3.26368,11.007431 L6.26591,5.75557323 C6.50872639,5.33080991 6.94150028,5.05390285 7.42253384,5.00706655 L7.56815,5 L10.1376351,5 Z"
-        fill={fill}
-      />
-      <path
-        d="M12,12 L8.14063365,19 L6.86158984,19 C6.58544746,19 6.36158984,18.7761424 6.36158984,18.5 C6.36158984,18.4129865 6.3842975,18.3274794 6.42746827,18.2519305 L9.98799988,12.02 L12,12 Z M10.1376351,5 C10.2246769,5 10.310211,5.02272249 10.3857773,5.06592003 C10.5988758,5.18773797 10.6883289,5.44260963 10.6093253,5.66611793 L10.571715,5.74814226 L7.28153572,11.5037155 C7.12772231,11.7727842 7.10849564,12.0954832 7.22385569,12.3779595 L7.28153572,12.4962845 L7.63499988,13.115 L5.48399988,16.878 L3.26368,12.992569 C2.9414043,12.428806 2.91454799,11.7473239 3.18311107,11.163941 L3.26368,11.007431 L6.26591,5.75557323 C6.50872639,5.33080991 6.94150028,5.05390285 7.42253384,5.00706655 L7.56815,5 L10.1376351,5 Z M15.8593663,5 L12,12 L9.999,12 L13.7120794,5.50386106 C13.8901222,5.19228606 14.2214657,5 14.5803225,5 L15.8593663,5 Z"
-        fillOpacity="0.2"
-        fill="var(--theme-rank-highlight)"
-      />
-      <path
-        d="M13.8580329,19 C13.7709911,19 13.685457,18.9772775 13.6098907,18.93408 C13.3967922,18.812262 13.3073391,18.5573904 13.3863427,18.3338821 L13.423953,18.2518577 L16.7141323,12.4962845 C16.8066247,12.3344857 16.8504511,12.1532944 16.8456114,11.9731159 L20.9954876,11.9731159 C20.9994405,12.2674994 20.9384636,12.5625597 20.8125569,12.836059 L20.731988,12.992569 L17.729758,18.2444268 C17.4869416,18.6691901 17.0541677,18.9460972 16.5731342,18.9929334 L16.427518,19 L13.8580329,19 Z M10.1376351,5 C10.2246769,5 10.310211,5.02272249 10.3857773,5.06592003 C10.5988758,5.18773797 10.6883289,5.44260963 10.6093253,5.66611793 L10.571715,5.74814226 L7.28153572,11.5037155 C7.18904327,11.6655143 7.14521691,11.8467056 7.15005663,12.0268841 L3.00018036,12.0268841 C2.99622748,11.7325006 3.05720438,11.4374403 3.18311107,11.163941 L3.26368,11.007431 L6.26591,5.75557323 C6.50872639,5.33080991 6.94150028,5.05390285 7.42253384,5.00706655 L7.56815,5 L10.1376351,5 Z"
-        fillOpacity="0.2"
-        fill="var(--theme-rank-highlight)"
-      />
-      <path
-        d="M15.8593663,5 L11.8555397,8.75193053 L13.7120794,5.50386106 C13.8901222,5.19228606 14.2214657,5 14.5803225,5 L15.8593663,5 Z"
-        fillOpacity="0.64"
-        fill="#FFFFFF"
-      />
-    </>
-  ),
-  (fill) => (
-    <>
-      <path
-        d="M18.5119999,7.125 L20.731988,11.007431 C21.0835615,11.6224452 21.0835615,12.3775548 20.731988,12.992569 L17.729758,18.2444268 C17.46266,18.7116664 16.9657134,19 16.427518,19 L13.8580333,19 C13.5818909,19 13.3580333,18.7761424 13.3580333,18.5 C13.3580333,18.4129581 13.3807558,18.327424 13.4239533,18.2518577 L16.7141326,12.4962845 C16.8899194,12.1887774 16.8899194,11.8112226 16.7141326,11.5037155 L16.3619999,10.888 L18.5119999,7.125 Z M8.50699988,14.611 L9.0669873,15.580127 C9.31057747,16.0020376 9.7419704,16.2766727 10.2210207,16.3231195 L10.3660254,16.330127 L11.5249999,16.33 L10.2879206,18.4961389 C10.1098778,18.8077139 9.77853431,19 9.41967746,19 L6.86158984,19 C6.58544746,19 6.36158984,18.7761424 6.36158984,18.5 C6.36158984,18.4129865 6.3842975,18.3274794 6.42746827,18.2519305 L8.50699988,14.611 Z M10.1376351,5 C10.2246769,5 10.310211,5.02272249 10.3857773,5.06592003 C10.5988758,5.18773797 10.6883289,5.44260963 10.6093253,5.66611793 L10.571715,5.74814226 L7.28153572,11.5037155 C7.12772231,11.7727842 7.10849564,12.0954832 7.22385569,12.3779595 L7.28153572,12.4962845 L7.63499988,13.115 L5.48399988,16.878 L3.26368,12.992569 C2.9414043,12.428806 2.91454799,11.7473239 3.18311107,11.163941 L3.26368,11.007431 L6.26591,5.75557323 C6.50872639,5.33080991 6.94150028,5.05390285 7.42253384,5.00706655 L7.56815,5 L10.1376351,5 Z M13.3452995,9.16987298 L13.4331972,9.17764841 C13.5481189,9.19816128 13.6524041,9.25837033 13.7276296,9.347639 L13.7783122,9.41987298 L15.1236116,11.75 L15.1608268,11.8300094 C15.200523,11.939791 15.200523,12.060209 15.1608268,12.1699906 L15.1236116,12.25 L13.7783122,14.580127 L13.7276296,14.652361 C13.6524041,14.7416297 13.5481189,14.8018387 13.4331972,14.8223516 L13.3452995,14.830127 L10.6547005,14.830127 L10.5668028,14.8223516 C10.4518811,14.8018387 10.3475959,14.7416297 10.2723704,14.652361 L10.2216878,14.580127 L8.87638837,12.25 L8.83917321,12.1699906 C8.79947703,12.060209 8.79947703,11.939791 8.83917321,11.8300094 L8.87638837,11.75 L10.2216878,9.41987298 L10.2723704,9.347639 C10.3475959,9.25837033 10.4518811,9.19816128 10.5668028,9.17764841 L10.6547005,9.16987298 L13.3452995,9.16987298 Z M17.1384102,5 C17.4145525,5 17.6384102,5.22385763 17.6384102,5.5 C17.6384102,5.58701347 17.6157025,5.67252062 17.5725317,5.74806947 L15.4919999,9.388 L14.9330127,8.41987298 C14.6894225,7.99796242 14.2580296,7.72332726 13.7789793,7.67688046 L13.6339746,7.66987298 L12.4739999,7.669 L13.7120794,5.50386106 C13.8901222,5.19228606 14.2214657,5 14.5803225,5 L17.1384102,5 Z"
-        fill={fill}
-      />
-      <path
-        d="M8.50658984,14.611 L9.0669873,15.580127 C9.21720124,15.8403052 9.4388324,16.0444778 9.69901603,16.1736704 L8.14158984,18.999 L6.86158984,19 C6.58544746,19 6.36158984,18.7761424 6.36158984,18.5 C6.36158984,18.4129865 6.3842975,18.3274794 6.42746827,18.2519305 L8.50658984,14.611 Z M10.1376351,5 C10.2246769,5 10.310211,5.02272249 10.3857773,5.06592003 C10.5988758,5.18773797 10.6883289,5.44260963 10.6093253,5.66611793 L10.571715,5.74814226 L7.28153572,11.5037155 C7.18904327,11.6655143 7.14521691,11.8467056 7.15005663,12.0268841 L3.00018036,12.0268841 C2.99622748,11.7325006 3.05720438,11.4374403 3.18311107,11.163941 L3.26368,11.007431 L6.26591,5.75557323 C6.50872639,5.33080991 6.94150028,5.05390285 7.42253384,5.00706655 L7.56815,5 L10.1376351,5 Z M15.8593663,5 L14.300984,7.82632959 C14.1392482,7.74602068 13.9626152,7.69468507 13.7789793,7.67688046 L13.6339746,7.66987298 L12.4745898,7.669 L13.7120794,5.50386106 C13.8901222,5.19228606 14.2214657,5 14.5803225,5 L15.8593663,5 Z"
-        id="Shine"
-        fillOpacity="0.2"
-        fill="var(--theme-rank-highlight)"
-      />
-      <path
-        d="M20.9954876,11.9731159 C20.9994405,12.2674994 20.9384636,12.5625597 20.8125569,12.836059 L20.731988,12.992569 L17.729758,18.2444268 C17.4869416,18.6691901 17.0541677,18.9460972 16.5731342,18.9929334 L16.427518,19 L13.8580329,19 C13.7709911,19 13.685457,18.9772775 13.6098907,18.93408 C13.3967922,18.812262 13.3073391,18.5573904 13.3863427,18.3338821 L13.423953,18.2518577 L16.7141323,12.4962845 C16.8066247,12.3344857 16.8504511,12.1532944 16.8456114,11.9731159 L20.9954876,11.9731159 Z M13.3452995,9.16987298 L13.4331972,9.17764841 C13.5030188,9.19011117 13.5689142,9.21722686 13.6267517,9.25661006 L12.029888,12.0182369 L15.1901573,12.021023 C15.1879362,12.0738639 15.1773372,12.1264313 15.1583603,12.1766627 L15.1236116,12.25 L13.7783122,14.580127 L13.7276296,14.652361 C13.6933923,14.6929897 13.6531356,14.7275989 13.6085963,14.7551859 L12.0344003,12.0260523 L10.4411564,14.7822341 C10.3732628,14.7501585 10.3128707,14.7029349 10.2650839,14.6435099 L10.2216878,14.580127 L8.87638837,12.25 L8.83917321,12.1699906 C8.82212063,12.1228309 8.81239345,12.0737084 8.80999168,12.0243094 L12.0249598,12.025701 L10.4153031,9.2309094 C10.463458,9.20464532 10.5159728,9.18627272 10.5708537,9.17694255 L10.6547005,9.16987298 L13.3452995,9.16987298 Z M10.1376351,5 C10.2246769,5 10.310211,5.02272249 10.3857773,5.06592003 C10.5988758,5.18773797 10.6883289,5.44260963 10.6093253,5.66611793 L10.571715,5.74814226 L7.28153572,11.5037155 C7.18904327,11.6655143 7.14521691,11.8467056 7.15005663,12.0268841 L3.00018036,12.0268841 C2.99622748,11.7325006 3.05720438,11.4374403 3.18311107,11.163941 L3.26368,11.007431 L6.26591,5.75557323 C6.50872639,5.33080991 6.94150028,5.05390285 7.42253384,5.00706655 L7.56815,5 L10.1376351,5 Z"
-        id="Light"
-        fillOpacity="0.2"
-        fill="var(--theme-rank-highlight)"
-      />
-      <path
-        d="M10.2216878,9.41987298 L10.2723704,9.347639 C10.3131617,9.29923275 10.3624977,9.25937116 10.4174409,9.22975015 L12.0304099,12.0294249 L8.81007095,12.0258885 C8.80679602,11.9626494 8.81552917,11.8989277 8.83627043,11.8382626 L8.87638837,11.75 L10.2216878,9.41987298 Z"
-        id="Blink"
-        fillOpacity="0.64"
-        fill="#FFFFFF"
-      />
-    </>
-  ),
-  (fill) => (
-    <>
-      <path
-        d="M12.2207603,13.5518602 L12.3193877,13.7369958 L15.219794,18.7875116 C15.3571965,19.0267725 15.2748225,19.3321116 15.0357267,19.4698013 L12.7066821,20.8110431 L12.4805172,20.9160192 C12.1701946,21.0279936 11.8298054,21.0279936 11.5194828,20.9160192 L11.2933179,20.8110431 L8.96126973,19.4678025 C8.72220631,19.3301039 8.63984888,19.0247945 8.77723146,18.7855494 L11.67628,13.7369958 L11.7747767,13.5519915 L11.7747767,13.5519915 C11.8356309,13.4288186 11.9848143,13.3782993 12.1079872,13.4391535 C12.1569258,13.4633319 12.1965531,13.5029359 12.2207603,13.5518602 Z M20.4955826,5 C20.5826245,5 20.6681586,5.02272249 20.7437249,5.06592003 C20.9568233,5.18773797 21.0462765,5.44260963 20.9672729,5.66611793 L20.9296626,5.74814226 L17.6394833,11.5037155 C17.4856699,11.7727842 17.4664432,12.0954832 17.5818033,12.3779595 L17.6394833,12.4962845 L20.9296626,18.2518577 C20.9728602,18.327424 20.9955826,18.4129581 20.9955826,18.5 C20.9955826,18.7454599 20.8187075,18.9496084 20.5854583,18.9919443 L20.4955826,19 L17.9260976,19 C17.4368291,19 16.9816504,18.7617078 16.7022595,18.3673379 L16.6238576,18.2444268 L13.6216276,12.992569 C13.2993519,12.428806 13.2724956,11.7473239 13.5410587,11.163941 L13.6216276,11.007431 L16.6238576,5.75557323 C16.866674,5.33080991 17.2994479,5.05390285 17.7804814,5.00706655 L17.9260976,5 L20.4955826,5 Z M6.06957008,5 C6.60776544,5 7.10471206,5.28833357 7.37181008,5.75557323 L10.3740401,11.007431 C10.7256136,11.6224452 10.7256136,12.3775548 10.3740401,12.992569 L7.37181008,18.2444268 C7.10471206,18.7116664 6.60776544,19 6.06957008,19 L3.50008536,19 C3.22394299,19 3.00008536,18.7761424 3.00008536,18.5 C3.00008536,18.4129581 3.02280785,18.327424 3.06600539,18.2518577 L6.3561847,12.4962845 C6.53197145,12.1887774 6.53197145,11.8112226 6.3561847,11.5037155 L3.06600539,5.74814226 C2.92896021,5.50840651 3.01220736,5.20296522 3.2519431,5.06592003 C3.32750939,5.02272249 3.41304347,5 3.50008536,5 L6.06957008,5 Z M12.4805172,3.08398084 L12.7066821,3.1889569 L15.0357267,4.53019876 C15.2748225,4.6678884 15.3571965,4.97322748 15.219794,5.21248838 L12.3193877,10.2630042 L12.2208909,10.4480085 L12.2208909,10.4480085 C12.1600368,10.5711814 12.0108534,10.6217007 11.8876805,10.5608465 C11.8550547,10.5447276 11.8265674,10.5217528 11.8040283,10.4937312 L11.7748299,10.4480085 L11.7748299,10.4480085 L11.67628,10.2630042 L8.77723145,5.21445059 C8.63984887,4.97520544 8.7222063,4.66989606 8.96126972,4.5321975 L11.2933179,3.1889569 L11.5194828,3.08398084 C11.8298054,2.97200639 12.1701946,2.97200639 12.4805172,3.08398084 Z"
-        fill={fill}
-      />
-      <path
-        d="M12.0004314,13.4133638 L12,21 C11.837322,21 11.6746441,20.9720064 11.5194828,20.9160192 L11.2933179,20.8110431 L8.96126973,19.4678025 C8.72220631,19.3301039 8.63984888,19.0247945 8.77723146,18.7855494 L11.67628,13.7369958 L11.7747767,13.5519915 C11.8185157,13.463461 11.9078856,13.4124635 12.0004314,13.4133638 Z M6.06957008,5 C6.60776544,5 7.10471206,5.28833357 7.37181008,5.75557323 L10.3740401,11.007431 C10.7256136,11.6224452 10.7256136,12.3775548 10.3740401,12.992569 L7.37181008,18.2444268 C7.10471206,18.7116664 6.60776544,19 6.06957008,19 L3.50008536,19 C3.22394299,19 3.00008536,18.7761424 3.00008536,18.5 C3.00008536,18.4129581 3.02280785,18.327424 3.06600539,18.2518577 L6.3561847,12.4962845 C6.53197145,12.1887774 6.53197145,11.8112226 6.3561847,11.5037155 L3.06600539,5.74814226 C2.92896021,5.50840651 3.01220736,5.20296522 3.2519431,5.06592003 C3.32750939,5.02272249 3.41304347,5 3.50008536,5 L6.06957008,5 Z M12,3 L12.0004944,10.5866362 C11.9625825,10.587051 11.9240007,10.5787907 11.8876805,10.5608465 C11.8550547,10.5447276 11.8265674,10.5217528 11.8040283,10.4937312 L11.7748299,10.4480085 L11.67628,10.2630042 L8.77723145,5.21445059 C8.63984887,4.97520544 8.7222063,4.66989606 8.96126972,4.5321975 L11.2933179,3.1889569 L11.5194828,3.08398084 C11.6229237,3.04665602 11.7297053,3.02177281 11.8376004,3.0093312 L12,3 Z"
-        fillOpacity="0.2"
-        fill="var(--theme-rank-highlight)"
-      />
-      <path
-        d="M10.3660931,12.9770771 C10.374508,12.9754206 10.3784904,12.9771635 10.3780744,12.9823058 L10.3740401,12.992569 L7.37181008,18.2444268 C7.1289937,18.6691901 6.6962198,18.9460972 6.21518624,18.9929334 L6.06957008,19 L3.50008536,19 C3.41304347,19 3.32750939,18.9772775 3.2519431,18.93408 C3.03884466,18.812262 2.94939153,18.5573904 3.02839512,18.3338821 L3.06600539,18.2518577 L3.97156273,17.5789548 L7.60173907,14.9028347 L9.06175661,13.8478786 L9.82498374,13.3129136 L10.1148363,13.1194021 L10.2624116,13.0278104 L10.3275503,12.9922453 L10.3660931,12.9770771 Z M13.6294894,12.9770771 L13.6680323,12.9922453 L13.733171,13.0278104 L14.0123746,13.2059594 L14.6825162,13.6695354 L16.5893124,15.0455482 L20.0240198,17.5789548 L20.9295771,18.2518577 C21.0666223,18.4915935 20.9833752,18.7970348 20.7436394,18.93408 C20.6680731,18.9772775 20.582539,19 20.4954972,19 L17.9260124,19 C17.3878171,19 16.8908705,18.7116664 16.6237724,18.2444268 L13.6215424,12.992569 C13.614218,12.9797562 13.616867,12.9745923 13.6294894,12.9770771 Z M20.4954972,5 C20.582539,5 20.6680731,5.02272249 20.7436394,5.06592003 C20.9567379,5.18773797 21.046191,5.44260963 20.9671874,5.66611793 L20.9295771,5.74814226 L20.0240198,6.42104524 L16.5893124,8.95445182 L14.6825162,10.3304646 L14.0123746,10.7940406 L13.733171,10.9721896 L13.6680323,11.0077547 L13.6294894,11.0229229 C13.6210745,11.0245794 13.6170921,11.0228365 13.6175082,11.0176942 L13.6215424,11.007431 L16.6237724,5.75557323 C16.8665888,5.33080991 17.2993627,5.05390285 17.7803963,5.00706655 L17.9260124,5 L20.4954972,5 Z M6.06957008,5 C6.60776544,5 7.10471206,5.28833357 7.37181008,5.75557323 L10.3740401,11.007431 C10.3813645,11.0202438 10.3787155,11.0254077 10.3660931,11.0229229 L10.3275503,11.0077547 L10.2624116,10.9721896 L10.1148363,10.8805979 L9.82498374,10.6870864 L9.06175661,10.1521214 L7.60173907,9.09716527 L3.97156273,6.42104524 L3.06600539,5.74814226 C2.92896021,5.50840651 3.01220736,5.20296522 3.2519431,5.06592003 C3.32750939,5.02272249 3.41304347,5 3.50008536,5 L6.06957008,5 Z"
-        fillOpacity="0.2"
-        fill="var(--theme-rank-highlight)"
-      />
-      <path
-        d="M12,3 L12.0002626,10.5866362 C11.9624253,10.5870173 11.9239267,10.5787541 11.8876805,10.5608465 C11.8550547,10.5447276 11.8265674,10.5217528 11.8040283,10.4937312 L11.7748299,10.4480085 L11.67628,10.2630042 L8.77723145,5.21445059 C8.63984887,4.97520544 8.7222063,4.66989606 8.96126972,4.5321975 L11.2933179,3.1889569 L11.5194828,3.08398084 C11.6746441,3.02799361 11.837322,3 12,3 Z"
-        fillOpacity="0.64"
-        fill="#FFFFFF"
-      />
-    </>
-  ),
-  (fill) => (
-    <>
-      <path
-        d="M10.0289999,16.618 L11.4998396,17.4654261 C11.6543261,17.5544281 11.8266648,17.5989322 11.9990049,17.5989377 L12.1279813,17.5905983 C12.2564002,17.5739154 12.3823096,17.532202 12.4981787,17.4654579 L13.9669999,16.619 L15.2185617,18.787529 C15.3416265,18.9999102 15.2905229,19.2651143 15.1096724,19.4183728 L15.035207,19.4701006 L15.035207,19.4701006 L12.7066821,20.8110431 L12.4805172,20.9160192 C12.1701946,21.0279936 11.8298054,21.0279936 11.5194828,20.9160192 L11.2933179,20.8110431 L8.96184182,19.468132 C8.74898126,19.3458984 8.660343,19.0907671 8.74007405,18.8675214 L8.77844751,18.7854876 L8.77844751,18.7854876 L10.0289999,16.618 Z M17.7839999,12.75 L20.9296626,18.2518577 C20.9728602,18.327424 20.9955826,18.4129581 20.9955826,18.5 C20.9955826,18.7454599 20.8187075,18.9496084 20.5854583,18.9919443 L20.4955826,19 L17.9260976,19 C17.4368291,19 16.9816504,18.7617078 16.7022595,18.3673379 L16.6238576,18.2444268 L15.2659999,15.87 L16.496142,15.1625094 C16.8060265,14.9840065 16.997,14.6536087 16.997,14.2959892 L16.9969999,12.75 L17.7839999,12.75 Z M6.99999988,12.75 L7,14.2950317 C7,14.6526284 7.19094941,14.9830089 7.50080283,15.1615201 L8.72899988,15.869 L7.37181008,18.2444268 C7.10471206,18.7116664 6.60776544,19 6.06957008,19 L3.50008536,19 C3.22394299,19 3.00008536,18.7761424 3.00008536,18.5 C3.00008536,18.4129581 3.02280785,18.327424 3.06600539,18.2518577 L6.21099988,12.75 L6.99999988,12.75 Z M11.4999115,8.26347498 C11.770231,8.10777096 12.095196,8.08829946 12.3791838,8.20506326 L12.4981053,8.26344583 L14.9950716,9.7015042 C15.2705597,9.86016394 15.4520865,10.1388654 15.4890246,10.4500283 L15.496,10.568065 L15.496,13.428935 C15.496,13.7468448 15.3450831,14.0432408 15.0938763,14.2305425 L14.9950716,14.2954958 L12.4981053,15.7335542 C12.2277767,15.8892424 11.9028106,15.9086949 11.6188296,15.7919145 L11.4999115,15.733525 L8.99987779,14.2935056 C8.724415,14.134839 8.54290855,13.8561494 8.50597455,13.5450039 L8.499,13.426974 L8.499,10.570026 C8.499,10.2521348 8.64989948,9.9557535 8.90108231,9.76844941 L8.99987779,9.70349439 L11.4999115,8.26347498 Z M6.06957008,5 C6.60776544,5 7.10471206,5.28833357 7.37181008,5.75557323 L8.72799988,8.128 L7.50080283,8.83647987 C7.19094941,9.01499114 6.99999988,9.34537157 6.99999988,9.70296829 L6.99999988,11.25 L6.21099988,11.25 L3.06600539,5.74814226 C2.92896021,5.50840651 3.01220736,5.20296522 3.2519431,5.06592003 C3.32750939,5.02272249 3.41304347,5 3.50008536,5 L6.06957008,5 Z M20.4955826,5 C20.5826245,5 20.6681586,5.02272249 20.7437249,5.06592003 C20.9568233,5.18773797 21.0462765,5.44260963 20.9672729,5.66611793 L20.9296626,5.74814226 L17.7839999,11.25 L16.9969999,11.25 L16.997,9.70201076 C16.997,9.34439129 16.8060265,9.01399345 16.496142,8.83549056 L15.2669999,8.127 L16.6238576,5.75557323 C16.866674,5.33080991 17.2994479,5.05390285 17.7804814,5.00706655 L17.9260976,5 L20.4955826,5 Z M12.4805172,3.08398084 L12.7066821,3.1889569 L15.035207,4.52989946 C15.2481385,4.65200971 15.3368056,4.90712172 15.2570946,5.130373 L15.2185424,5.21245988 L15.2185424,5.21245988 L13.9679999,7.379 L12.4981787,6.53254213 C12.3823096,6.46579798 12.2564002,6.42408459 12.1279813,6.40740172 L11.9990049,6.39906234 C11.8266648,6.39906783 11.6543261,6.44357188 11.4998396,6.53257392 L10.0279999,7.38 L8.77846674,5.21450128 C8.6554931,5.00206739 8.70663027,4.73686378 8.88749649,4.58362226 L8.96184181,4.53186798 L8.96184181,4.53186798 L11.2933179,3.1889569 L11.5194828,3.08398084 C11.8298054,2.97200639 12.1701946,2.97200639 12.4805172,3.08398084 Z"
-        fill={fill}
-      />
-      <path
-        d="M10.0289999,16.618 L11.4998396,17.4654261 C11.6157045,17.5321776 11.7416112,17.573899 11.870029,17.5905901 L12,17.598 L12,21 C11.837322,21 11.6746441,20.9720064 11.5194828,20.9160192 L11.2933179,20.8110431 L8.96184182,19.468132 C8.74898126,19.3458984 8.660343,19.0907671 8.74007405,18.8675214 L8.77844751,18.7854876 L10.0289999,16.618 Z M6.99999988,12.75 L7,14.2950317 C7,14.6526284 7.19094941,14.9830089 7.50080283,15.1615201 L8.72899988,15.869 L7.37181008,18.2444268 C7.10471206,18.7116664 6.60776544,19 6.06957008,19 L3.50008536,19 C3.22394299,19 3.00008536,18.7761424 3.00008536,18.5 C3.00008536,18.4129581 3.02280785,18.327424 3.06600539,18.2518577 L6.21099988,12.75 L6.99999988,12.75 Z M12.0040527,8.13001916 C12.1316035,8.13065859 12.2590595,8.15567321 12.3791838,8.20506326 L12.4981053,8.26344583 L14.9950716,9.7015042 C15.1483119,9.78975868 15.2724792,9.91515577 15.35886,10.0626219 L12.0079999,11.995 L15.35886,13.9343781 C15.2918906,14.0487058 15.2022092,14.1497686 15.0938763,14.2305425 L14.9950716,14.2954958 L12.4981053,15.7335542 C12.3457438,15.8213026 12.1760259,15.8657739 12.006072,15.8669687 L12.008,12.0040245 L8.64259934,13.9432862 C8.57002908,13.8229088 8.52287185,13.687353 8.50597455,13.5450039 L8.499,13.426974 L8.499,10.570026 C8.499,10.3860464 8.54954402,10.2092717 8.64081995,10.0566769 L11.9999999,11.999 L12.0040527,8.13001916 Z M6.06957008,5 C6.60776544,5 7.10471206,5.28833357 7.37181008,5.75557323 L8.72799988,8.128 L7.50080283,8.83647987 C7.19094941,9.01499114 6.99999988,9.34537157 6.99999988,9.70296829 L6.99999988,11.25 L6.21099988,11.25 L3.06600539,5.74814226 C2.92896021,5.50840651 3.01220736,5.20296522 3.2519431,5.06592003 C3.32750939,5.02272249 3.41304347,5 3.50008536,5 L6.06957008,5 Z M12,3 L11.9990049,6.39906234 C11.8266648,6.39906783 11.6543261,6.44357188 11.4998396,6.53257392 L10.0279999,7.38 L8.77846674,5.21450128 C8.6554931,5.00206739 8.70663027,4.73686378 8.88749649,4.58362226 L8.96184181,4.53186798 L11.2933179,3.1889569 L11.5194828,3.08398084 C11.6746441,3.02799361 11.837322,3 12,3 Z"
-        fillOpacity="0.2"
-        fill="var(--theme-rank-highlight)"
-      />
-      <path
-        d="M8.98199988,18.43 L11.2933179,19.3127665 L11.5194828,19.4177426 C11.8298054,19.529717 12.1701946,19.529717 12.4805172,19.4177426 L12.7066821,19.3127665 L15.0119999,18.431 L15.2185617,18.787529 C15.3416265,18.9999102 15.2905229,19.2651143 15.1096724,19.4183728 L15.035207,19.4701006 L12.7066821,20.8110431 L12.4805172,20.9160192 C12.3770763,20.953344 12.2702947,20.9782272 12.1623996,20.9906688 L12,21 C11.837322,21 11.6746441,20.9720064 11.5194828,20.9160192 L11.2933179,20.8110431 L8.96184182,19.468132 C8.74898126,19.3458984 8.660343,19.0907671 8.74007405,18.8675214 L8.77844751,18.7854876 L8.98199988,18.43 Z M19.9519999,16.543 L20.9296626,18.2518577 C20.9728602,18.327424 20.9955826,18.4129581 20.9955826,18.5 C20.9955826,18.7454599 20.8187075,18.9496084 20.5854583,18.9919443 L20.4955826,19 L17.9260976,19 C17.4368291,19 16.9816504,18.7617078 16.7022595,18.3673379 L16.6238576,18.2444268 L16.4219999,17.892 L19.9519999,16.543 Z M4.04099988,16.545 L7.57199988,17.892 L7.37181008,18.2444268 C7.10471206,18.7116664 6.60776544,19 6.06957008,19 L3.50008536,19 C3.22394299,19 3.00008536,18.7761424 3.00008536,18.5 C3.00008536,18.4129581 3.02280785,18.327424 3.06600539,18.2518577 L4.04099988,16.545 Z M20.4955826,5 C20.5826245,5 20.6681586,5.02272249 20.7437249,5.06592003 C20.9568233,5.18773797 21.0462765,5.44260963 20.9672729,5.66611793 L20.9296626,5.74814226 L19.9519999,7.457 L16.4209999,6.108 L16.6238576,5.75557323 C16.866674,5.33080991 17.2994479,5.05390285 17.7804814,5.00706655 L17.9260976,5 L20.4955826,5 Z M6.06957008,5 C6.60776544,5 7.10471206,5.28833357 7.37181008,5.75557323 L7.57299988,6.108 L4.04199988,7.456 L3.06600539,5.74814226 C2.92896021,5.50840651 3.01220736,5.20296522 3.2519431,5.06592003 C3.32750939,5.02272249 3.41304347,5 3.50008536,5 L6.06957008,5 Z M12.4805172,3.08398084 L12.7066821,3.1889569 L15.035207,4.52989946 C15.2481385,4.65200971 15.3368056,4.90712172 15.2570946,5.130373 L15.2185424,5.21245988 L15.0119999,5.569 L12.7066821,4.6889569 L12.4805172,4.58398084 C12.2219151,4.4906688 11.9424333,4.47511679 11.6768712,4.53732482 L11.5194828,4.58398084 L11.2933179,4.6889569 L8.98299988,5.57 L8.77846674,5.21450128 C8.6554931,5.00206739 8.70663027,4.73686378 8.88749649,4.58362226 L8.96184181,4.53186798 L11.2933179,3.1889569 L11.5194828,3.08398084 C11.8298054,2.97200639 12.1701946,2.97200639 12.4805172,3.08398084 Z"
-        fillOpacity="0.2"
-        fill="var(--theme-rank-highlight)"
-      />
-      <path
-        d="M12.0050154,8.13000658 L12.0028579,12.0049501 L8.6402291,10.057666 C8.70673402,9.94619233 8.79495756,9.84758539 8.90108231,9.76844941 L8.99987779,9.70349439 L11.4999115,8.26347498 C11.6561759,8.17346668 11.8307011,8.12898299 12.0050154,8.13000658 Z"
-        fillOpacity="0.64"
-        fill="#FFFFFF"
-      />
-    </>
-  ),
-  (fill) => (
-    <>
-      <path
-        d="M12.0027158,17.949 L12.5531404,18.4134483 L13.0522401,18.8182988 L13.3045035,19.0172039 L13.8872893,19.4624402 L13.8872893,19.4624402 L14.1023462,19.620622 L14.1023462,19.620622 C14.2139993,19.7018464 14.2381933,19.8582778 14.1566318,19.9696841 C14.1361159,19.9977071 14.1099862,20.0211481 14.0791207,20.0371481 L12.7572702,20.798888 L12.6147667,20.8712883 C12.2272392,21.0429039 11.7841925,21.0429039 11.3966649,20.8712883 L11.2541614,20.798888 L9.93277621,20.0358859 C9.81320697,19.9668435 9.77224686,19.8139436 9.84128924,19.6943744 C9.85908368,19.6635575 9.88325202,19.6368946 9.91217842,19.6161683 L9.95289761,19.5869922 C10.1146557,19.4679163 10.2781343,19.3449892 10.4433016,19.2183238 L10.9438377,18.827226 L11.4433657,18.4221972 L12.0027158,17.949 Z M16.811,12.5 L16.8363648,12.5004028 L17.2454086,12.9767355 L17.4432018,13.2129576 L17.4432018,13.2129576 L17.9171745,13.7968494 L17.9171745,13.7968494 L18.3604682,14.3696751 L18.6918561,14.8186687 C18.7453103,14.8927526 18.7978636,14.9664486 18.8494978,15.0397361 L19.1481248,15.4743936 C20.5846365,17.6215188 21.1607641,19.3638338 20.3801568,20.144441 C19.4563961,21.0682018 17.1858541,20.0919833 14.4914343,18.038349 L13.9970744,17.6532717 L13.5086615,17.25717 C13.2360344,17.0318774 12.9603972,16.7971859 12.6826179,16.5538693 L13.3498943,16.9020104 C13.7392153,17.1066885 14.1851514,17.1773179 14.6186667,17.1029642 L14.7648536,17.0723253 C14.9489688,17.0265848 15.1214205,16.9561399 15.2792304,16.8652023 L15.4913848,17.0213228 L15.9050626,17.3154123 C15.9727683,17.3623716 16.039834,17.4082903 16.10623,17.4531489 L16.4963322,17.7094258 C17.8954998,18.6013315 18.9418207,18.9614568 19.3194967,18.5837808 C19.6922699,18.2110076 19.3462729,17.1868369 18.4793506,15.8146307 L18.2176546,15.4133635 C18.0798901,15.2082568 17.931879,14.9966369 17.7741981,14.7793883 L17.4461142,14.3376483 L17.0013098,13.766761 L17.0013098,13.766761 L16.5198197,13.1779485 L16.336,12.963 L16.811,12.5 Z M7.199,12.5 L7.671,12.961 L7.48737335,13.1779485 L7.09931049,13.6503343 C6.97376965,13.806037 6.85209923,13.9598933 6.73447013,14.111641 L6.3938777,14.5602963 C6.1751266,14.8548333 5.97322442,15.1398879 5.78953839,15.4133635 L5.52784239,15.8146307 C4.66092012,17.1868369 4.31492316,18.2110076 4.68769635,18.5837808 C4.99685672,18.8929412 5.75406497,18.7077131 6.78626825,18.1409571 L7.14028363,17.9380626 C7.2615255,17.865824 7.38592499,17.7890301 7.5132447,17.7078356 L7.90372717,17.4512051 C8.03664997,17.3613642 8.17225562,17.2672776 8.31030674,17.1691 L8.73155909,16.8624491 C9.3056545,17.1949426 10.0316536,17.2331599 10.6615373,16.9020104 L11.3196313,16.5582519 L10.995126,16.8392713 L10.4985315,17.25717 L10.0101187,17.6532717 L9.51575877,18.038349 L9.2720049,18.222052 L9.2720049,18.222052 L8.79211975,18.5711003 C6.41964354,20.2537753 4.46681871,20.9842235 3.62703618,20.144441 C2.8724492,19.389854 3.38565548,17.7366046 4.71830335,15.6877362 L5.00556892,15.2583524 C5.10512353,15.1134528 5.20842854,14.9668366 5.31533687,14.8186687 L5.6467248,14.3696751 L5.9988503,13.9123617 L6.37083097,13.4477185 C6.49803568,13.2917258 6.62840252,13.1346764 6.76178442,12.9767355 L7.17082824,12.5004028 L7.199,12.5 Z M12.2269987,7.68140769 C12.325605,7.73007281 12.4054188,7.80988657 12.4540839,7.90849288 L13.3588437,9.74173649 C13.3949036,9.81480179 13.4476212,9.87676538 13.5114281,9.92358146 C13.5764935,9.97132094 13.6530898,10.0033095 13.7353126,10.0152571 L15.758417,10.3092314 C16.0316894,10.3489402 16.2210302,10.6026617 16.1813214,10.8759341 C16.1655091,10.9847526 16.1142655,11.0853238 16.0355237,11.1620781 L14.5715916,12.5890593 C14.4537485,12.7039278 14.3999743,12.8694279 14.4277932,13.031625 L14.7733807,15.0465547 C14.8200612,15.3187229 14.6372673,15.5772009 14.3650991,15.6238814 C14.2567203,15.6424698 14.1452363,15.6248125 14.047906,15.5736429 L12.2383864,14.6223222 C12.0927244,14.5457431 11.9187073,14.5457431 11.7730453,14.6223222 L9.96352565,15.5736429 C9.71910336,15.7021433 9.41678978,15.6081701 9.28828937,15.3637478 C9.23711984,15.2664175 9.2194625,15.1549335 9.23805092,15.0465547 L9.58363842,13.031625 C9.61145739,12.8694279 9.55768315,12.7039278 9.4398401,12.5890593 L7.97590798,11.1620781 C7.77816619,10.9693275 7.77412004,10.6527709 7.96687064,10.4550291 C8.04362494,10.3762873 8.14419615,10.3250437 8.25301467,10.3092314 L10.2761191,10.0152571 C10.4389742,9.99159291 10.579757,9.88930822 10.652588,9.74173649 L11.5573478,7.90849288 C11.6795589,7.66086605 11.9793719,7.55919655 12.2269987,7.68140769 Z M20.005,10.0294865 L20.005,13.3048057 C20.0024696,13.4428537 19.8885407,13.552746 19.7504919,13.5502562 C19.6734939,13.5488676 19.6014328,13.5120698 19.555786,13.4500407 L19.4434214,13.3028452 L19.4434214,13.3028452 L19.0488288,12.8029288 L18.6317978,12.2937128 L18.22,11.808 L19.5494347,9.88720886 C19.6280125,9.7736783 19.7837471,9.74534346 19.8972776,9.82392122 C19.9647416,9.87061499 20.005,9.94743956 20.005,10.0294865 Z M4.4604396,9.90105643 L5.783,11.812 L5.37539526,12.2937128 L4.95836422,12.8029288 L4.5637716,13.3028452 C4.5303973,13.3459764 4.49335213,13.3943774 4.45263608,13.4480483 C4.36951286,13.5582966 4.21264606,13.5794845 4.10282528,13.4958002 C4.04001974,13.447942 4.00351567,13.3732254 4.00433499,13.2942681 L4.00487139,10.0432896 C4.00489358,9.90521843 4.11684038,9.7933298 4.25491156,9.7933298 C4.33694612,9.7933298 4.41375427,9.83360165 4.4604396,9.90105643 Z M4.02176289,4.5 C4.3088897,4.5 4.58002947,4.62325937 4.76845832,4.83481792 L4.84395508,4.93079002 L7.61601731,8.9345522 C7.34484653,9.03734524 7.09758825,9.19853031 6.8927411,9.40868188 L6.77754378,9.53737946 C6.6291597,9.71813478 6.51734305,9.91861107 6.44196948,10.1290797 L3.0889039,5.28460499 C3.03101398,5.20098621 3,5.10170223 3,5 C3,4.75454011 3.17687516,4.55039163 3.41012437,4.50805567 L3.5,4.5 L4.02176289,4.5 Z M20.5014317,4.5 C20.777574,4.5 21.0014317,4.72385763 21.0014317,5 C21.0014317,5.10170223 20.9704177,5.20098621 20.9125278,5.28460499 L17.5660392,10.1199064 L17.5660392,10.1199064 C17.3685501,9.57934694 16.9422777,9.13818776 16.3878339,8.93125141 L19.1574766,4.93079002 C19.34426,4.66099183 19.6515239,4.5 19.9796688,4.5 L20.5014317,4.5 Z M7.65176289,4.5 C7.9388897,4.5 8.21002947,4.62325937 8.39845832,4.83481792 L8.47395508,4.93079002 L10.155,7.36 L9.53971583,8.60667033 L9.066,8.675 L6.7189039,5.28460499 C6.66101398,5.20098621 6.63,5.10170223 6.63,5 C6.63,4.75454011 6.80687516,4.55039163 7.04012437,4.50805567 L7.13,4.5 L7.65176289,4.5 Z M16.8714317,4.5 C17.147574,4.5 17.3714317,4.72385763 17.3714317,5 C17.3714317,5.10170223 17.3404177,5.20098621 17.2825278,5.28460499 L14.935,8.674 L14.4707158,8.60667033 L13.851,7.351 L15.5274766,4.93079002 C15.71426,4.66099183 16.0215239,4.5 16.3496688,4.5 L16.8714317,4.5 Z M11.3966649,3.12871167 C11.7841925,2.95709611 12.2272392,2.95709611 12.6147667,3.12871167 L12.7572702,3.20111198 L14.0528229,3.94835555 C14.152492,4.00584227 14.1975998,4.12159001 14.1699459,4.22779037 L14.1391495,4.29863374 L14.1391495,4.29863374 L12.767,6.281 L12.7480254,6.2731973 C12.2468968,6.07227763 11.7050175,6.08889539 11.235227,6.28478564 L9.86953325,4.31220474 C9.79093938,4.19868533 9.81925213,4.04294675 9.93277155,3.96435288 L11.2541614,3.20111198 L11.3966649,3.12871167 Z"
-        fill={fill}
-      />
-      <path
-        d="M11.999,17.9510121 L11.9996875,20.9999879 C11.7940547,20.9991656 11.5885536,20.9562657 11.3966649,20.8712883 L11.2541614,20.798888 L9.93277621,20.0358859 C9.81320697,19.9668435 9.77224686,19.8139436 9.84128924,19.6943744 C9.85908368,19.6635575 9.88325202,19.6368946 9.91217842,19.6161683 L9.95289761,19.5869922 C10.1146557,19.4679163 10.2781343,19.3449892 10.4433016,19.2183238 L10.9438377,18.827226 L11.4433657,18.4221972 L11.999,17.9510121 Z M7.199,12.5 L7.671,12.961 L7.48737335,13.1779485 L7.09931049,13.6503343 C6.97376965,13.806037 6.85209923,13.9598933 6.73447013,14.111641 L6.3938777,14.5602963 C6.1751266,14.8548333 5.97322442,15.1398879 5.78953839,15.4133635 L5.52784239,15.8146307 C4.66092012,17.1868369 4.31492316,18.2110076 4.68769635,18.5837808 C4.99685672,18.8929412 5.75406497,18.7077131 6.78626825,18.1409571 L7.14028363,17.9380626 C7.2615255,17.865824 7.38592499,17.7890301 7.5132447,17.7078356 L7.90372717,17.4512051 C8.03664997,17.3613642 8.17225562,17.2672776 8.31030674,17.1691 L8.73155909,16.8624491 C9.3056545,17.1949426 10.0316536,17.2331599 10.6615373,16.9020104 L11.3196313,16.5582519 L10.995126,16.8392713 L10.4985315,17.25717 L10.0101187,17.6532717 L9.51575877,18.038349 L9.2720049,18.222052 L9.2720049,18.222052 L8.79211975,18.5711003 C6.41964354,20.2537753 4.46681871,20.9842235 3.62703618,20.144441 C2.8724492,19.389854 3.38565548,17.7366046 4.71830335,15.6877362 L5.00556892,15.2583524 C5.10512353,15.1134528 5.20842854,14.9668366 5.31533687,14.8186687 L5.6467248,14.3696751 L5.9988503,13.9123617 L6.37083097,13.4477185 C6.49803568,13.2917258 6.62840252,13.1346764 6.76178442,12.9767355 L7.17082824,12.5004028 L7.199,12.5 Z M4.25491156,9.7933298 C4.33694612,9.7933298 4.41375427,9.83360165 4.4604396,9.90105643 L5.783,11.812 L5.37539526,12.2937128 L4.95836422,12.8029288 L4.5637716,13.3028452 C4.5303973,13.3459764 4.49335213,13.3943774 4.45263608,13.4480483 C4.36951286,13.5582966 4.21264606,13.5794845 4.10282528,13.4958002 C4.04001974,13.447942 4.00351567,13.3732254 4.00433499,13.2942681 L4.00487139,10.0432896 C4.00489358,9.90521843 4.11684038,9.7933298 4.25491156,9.7933298 Z M4.02176289,4.5 C4.3088897,4.5 4.58002947,4.62325937 4.76845832,4.83481792 L4.84395508,4.93079002 L7.61601731,8.9345522 C7.34484653,9.03734524 7.09758825,9.19853031 6.8927411,9.40868188 L6.77754378,9.53737946 C6.6291597,9.71813478 6.51734305,9.91861107 6.44196948,10.1290797 L3.0889039,5.28460499 C3.03101398,5.20098621 3,5.10170223 3,5 C3,4.75454011 3.17687516,4.55039163 3.41012437,4.50805567 L3.5,4.5 L4.02176289,4.5 Z M7.65176289,4.5 C7.9388897,4.5 8.21002947,4.62325937 8.39845832,4.83481792 L8.47395508,4.93079002 L10.155,7.36 L9.53971583,8.60667033 L9.066,8.675 L6.7189039,5.28460499 C6.66101398,5.20098621 6.63,5.10170223 6.63,5 C6.63,4.75454011 6.80687516,4.55039163 7.04012437,4.50805567 L7.13,4.5 L7.65176289,4.5 Z M11.9996875,3.00001205 L11.9994072,6.13001535 C11.7358697,6.1310497 11.4760748,6.18435848 11.235227,6.28478564 L9.86953325,4.31220474 C9.79093938,4.19868533 9.81925213,4.04294675 9.93277155,3.96435288 L11.2541614,3.20111198 L11.3966649,3.12871167 C11.5885536,3.04373429 11.7940547,3.00083442 11.9996875,3.00001205 Z M12.2269987,7.68140769 C12.325605,7.73007281 12.4054188,7.80988657 12.4540839,7.90849288 L13.3588437,9.74173649 C13.3954459,9.81590074 13.4492108,9.87862682 13.5143147,9.92568361 L12.0094333,11.997 L16.1621452,10.6494712 C16.1852049,10.7202483 16.1927503,10.7972814 16.1813214,10.8759341 C16.1655091,10.9847526 16.1142655,11.0853238 16.0355237,11.1620781 L14.5715916,12.5890593 C14.5125405,12.6466198 14.4695772,12.7168939 14.4449066,12.7930952 L12.0104333,12.004 L14.5746352,15.5356027 C14.5144303,15.5794369 14.4434699,15.6104397 14.3650991,15.6238814 C14.2567203,15.6424698 14.1452363,15.6248125 14.047906,15.5736429 L12.2383864,14.6223222 C12.1652519,14.5838731 12.0849695,14.5647287 12.0047168,14.5648889 L12.0034333,12.004 L9.43718286,15.535883 C9.37680979,15.4921546 9.32532894,15.4342012 9.28828937,15.3637478 C9.23711984,15.2664175 9.2194625,15.1549335 9.23805092,15.0465547 L9.58363842,13.031625 C9.59763962,12.9499917 9.59097263,12.8675218 9.56587164,12.7910913 L11.9984333,12 L7.84921364,10.6496958 C7.87215149,10.5788139 7.91134578,10.5119917 7.96687064,10.4550291 C8.04362494,10.3762873 8.14419615,10.3250437 8.25301467,10.3092314 L10.2761191,10.0152571 C10.3575466,10.003425 10.4334561,9.9719378 10.4981126,9.92496213 L12.0034333,11.995 L12.0054645,7.62966382 C12.0799774,7.6295705 12.155654,7.64619699 12.2269987,7.68140769 Z"
-        fillOpacity="0.2"
-        fill="var(--theme-rank-highlight)"
-      />
-      <path
-        d="M4.02176289,4.5 C4.3088897,4.5 4.58002947,4.62325937 4.76845832,4.83481792 L4.84395508,4.93079002 L6.059,6.685 L4.476,7.29 L3.0889039,5.28460499 C3.03101398,5.20098621 3,5.10170223 3,5 C3,4.75454011 3.17687516,4.55039163 3.41012437,4.50805567 L3.5,4.5 L4.02176289,4.5 Z M20.5014317,4.5 C20.777574,4.5 21.0014317,4.72385763 21.0014317,5 C21.0014317,5.10170223 20.9704177,5.20098621 20.9125278,5.28460499 L19.5254312,7.28880423 L17.943,6.684 L19.1574766,4.93079002 C19.34426,4.66099183 19.6515239,4.5 19.9796688,4.5 L20.5014317,4.5 Z M7.65176289,4.5 C7.9388897,4.5 8.21002947,4.62325937 8.39845832,4.83481792 L8.47395508,4.93079002 L8.929,5.589 L7.347,6.193 L6.7189039,5.28460499 C6.66101398,5.20098621 6.63,5.10170223 6.63,5 C6.63,4.75454011 6.80687516,4.55039163 7.04012437,4.50805567 L7.13,4.5 L7.65176289,4.5 Z M16.8714317,4.5 C17.147574,4.5 17.3714317,4.72385763 17.3714317,5 C17.3714317,5.10170223 17.3404177,5.20098621 17.2825278,5.28460499 L16.654,6.191 L15.072,5.586 L15.5274766,4.93079002 C15.71426,4.66099183 16.0215239,4.5 16.3496688,4.5 L16.8714317,4.5 Z M12.6147667,3.12871167 L12.7572702,3.20111198 L14.0528229,3.94835555 C14.152492,4.00584227 14.1975998,4.12159001 14.1699459,4.22779037 L14.1391495,4.29863374 L13.629,5.035 L12.7572702,4.70111198 L12.6147667,4.62871167 C12.2756801,4.47854806 11.8940867,4.4597776 11.5443734,4.57240031 L11.3966649,4.62871167 L11.2541614,4.70111198 L10.372,5.038 L9.86953325,4.31220474 C9.79093938,4.19868533 9.81925213,4.04294675 9.93277155,3.96435288 L11.2541614,3.20111198 L11.3966649,3.12871167 C11.7841925,2.95709611 12.2272392,2.95709611 12.6147667,3.12871167 Z M10.7277468,18.996 L11.2541614,19.299888 L11.3966649,19.3722883 C11.7357516,19.5224519 12.117345,19.5412224 12.4670583,19.4285997 L12.6147667,19.3722883 L12.7572702,19.299888 L13.2797468,18.998 L13.8872893,19.4624402 L14.1023462,19.620622 C14.2139993,19.7018464 14.2381933,19.8582778 14.1566318,19.9696841 C14.1361159,19.9977071 14.1099862,20.0211481 14.0791207,20.0371481 L12.7572702,20.798888 L12.6147667,20.8712883 C12.2272392,21.0429039 11.7841925,21.0429039 11.3966649,20.8712883 L11.2541614,20.798888 L9.93277621,20.0358859 C9.81320697,19.9668435 9.77224686,19.8139436 9.84128924,19.6943744 C9.85908368,19.6635575 9.88325202,19.6368946 9.91217842,19.6161683 L9.95289761,19.5869922 C10.1146557,19.4679163 10.2781343,19.3449892 10.4433016,19.2183238 L10.7277468,18.996 Z M16.8363648,12.5004028 L17.2454086,12.9767355 L17.4432018,13.2129576 L17.8247791,13.6808944 L18.0083427,13.9123617 L18.3604682,14.3696751 L18.6918561,14.8186687 C18.7453103,14.8927526 18.7978636,14.9664486 18.8494978,15.0397361 L19.1481248,15.4743936 C20.5846365,17.6215188 21.1607641,19.3638338 20.3801568,20.144441 C19.4563961,21.0682018 17.1858541,20.0919833 14.4914343,18.038349 L14.6771446,18.1761785 C17.1809537,19.983945 18.4048229,18.8081525 17.7509445,18.413173 L17.6360598,18.3554527 C18.4551432,18.7496826 19.053347,18.8499305 19.3194967,18.5837808 C19.6922699,18.2110076 19.3462729,17.1868369 18.4793506,15.8146307 L18.2176546,15.4133635 C18.0798901,15.2082568 17.931879,14.9966369 17.7741981,14.7793883 L17.4461142,14.3376483 L17.0013098,13.766761 L16.5198197,13.1779485 L16.336,12.963 L16.811,12.5 L16.8363648,12.5004028 Z M7.199,12.5 L7.671,12.961 L7.48737335,13.1779485 L7.09931049,13.6503343 C6.97376965,13.806037 6.85209923,13.9598933 6.73447013,14.111641 L6.3938777,14.5602963 C6.1751266,14.8548333 5.97322442,15.1398879 5.78953839,15.4133635 L5.52784239,15.8146307 C4.66092012,17.1868369 4.31492316,18.2110076 4.68769635,18.5837808 C4.99685672,18.8929412 5.75406497,18.7077131 6.78626825,18.1409571 L6.70106535,18.192255 C5.90940047,18.7191519 7.79877865,19.1552077 8.7049189,18.6271519 L8.783,18.577 L8.55629665,18.7361708 C6.29069193,20.3007271 4.43882596,20.9562308 3.62703618,20.144441 C2.8724492,19.389854 3.38565548,17.7366046 4.71830335,15.6877362 L5.00556892,15.2583524 C5.10512353,15.1134528 5.20842854,14.9668366 5.31533687,14.8186687 L5.6467248,14.3696751 L5.9988503,13.9123617 L6.37083097,13.4477185 C6.49803568,13.2917258 6.62840252,13.1346764 6.76178442,12.9767355 L7.17082824,12.5004028 L7.199,12.5 Z M19.8972776,9.82392122 C19.9647416,9.87061499 20.005,9.94743956 20.005,10.0294865 L20.005,13.3048057 C20.0024696,13.4428537 19.8885407,13.552746 19.7504919,13.5502562 C19.6734939,13.5488676 19.6014328,13.5120698 19.555786,13.4500407 L19.5494347,9.88720886 C19.6280125,9.7736783 19.7837471,9.74534346 19.8972776,9.82392122 Z M4.25491156,9.7933298 C4.33694612,9.7933298 4.41375427,9.83360165 4.4604396,9.90105643 L4.4750716,12.4621492 L4.4750716,12.4621492 L4.46134768,13.3856363 L4.46134768,13.3856363 L4.45577556,13.4382512 C4.45476834,13.4433514 4.45372184,13.4466171 4.45263608,13.4480483 C4.36951286,13.5582966 4.21264606,13.5794845 4.10282528,13.4958002 C4.04001974,13.447942 4.00351567,13.3732254 4.00434896,13.2942681 L4.00487139,10.0432896 C4.00489358,9.90521843 4.11684038,9.7933298 4.25491156,9.7933298 Z"
-        fillOpacity="0.2"
-        fill="var(--theme-rank-highlight)"
-      />
-      <path
-        d="M12.0060521,7.62966422 L12.0074821,12.0054361 L10.4970907,9.92570256 C10.5622064,9.87864366 10.6159808,9.8159107 10.652588,9.74173649 L11.5573478,7.90849288 C11.6444408,7.73202318 11.8217308,7.62966422 12.0060521,7.62966422 Z"
-        fillOpacity="0.64"
-        fill="#FFFFFF"
-      />
-    </>
-  ),
-];
-
-const getColorPath = (rank, id) => {
-  return rankPaths[getRank(rank)](`url(#${id})`);
-};
-const getOutlinePath = (rank) => {
-  return outlinePaths[getRank(rank)];
-};
-
-export default forwardRef(function Rank(
-  { rank, colorByRank, style, className, ...props }: RankProps,
-  ref: LegacyRef<SVGSVGElement>,
-): ReactElement {
-  const [id] = useState(`rank-${Math.random().toString(36).substring(7)}`);
-
-  return (
-    <svg
-      {...props}
-      viewBox="0 0 24 24"
-      xmlns="http://www.w3.org/2000/svg"
-      ref={ref}
-      className={classNames(className)}
-      style={
-        colorByRank
-          ? ({
-              ...style,
-              '--stop-color1': `var(--theme-rank-${rank}-color-bottom)`,
-              '--stop-color2': `var(--theme-rank-${rank}-color-top)`,
-            } as CSSProperties)
-          : style
-      }
-    >
-      <defs>
-        <linearGradient x1="50%" y1="100%" x2="50%" y2="0%" id={id}>
-          <stop stopColor="var(--stop-color1)" offset="0%" />
-          <stop stopColor="var(--stop-color2)" offset="100%" />
-        </linearGradient>
-      </defs>
-      <g stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-        {colorByRank ? getColorPath(rank, id) : getOutlinePath(rank)}
-      </g>
-    </svg>
-  );
-});
