--- packages/shared/src/components/modals/RanksModal/index.module.css
@@ -1,22 +1,0 @@
-div.ranksModal {
-  & :global(.modal) {
-    align-items: center;
-    max-width: 26.25rem;
-  }
-
-  @screen mobileL {
-    & :global(.overlay) {
-      justify-content: flex-start;
-    }
-
-    & :global(.modal) {
-      margin-top: $headerHeight;
-      overflow: visible;
-    }
-  }
-}
-
-.rankProgress {
-  width: $headerRankHeight;
-  height: $headerRankHeight;
-}
