--- packages/shared/src/components/profile/ProfileTooltipContent.tsx
@@ -1,64 +1,0 @@
-import React, { ReactElement } from 'react';
-import { Author } from '../../graphql/comments';
-import Rank from '../Rank';
-import { ProfileImageLink } from './ProfileImageLink';
-import { ProfileLink } from './ProfileLink';
-import { TagLinks } from '../TagLinks';
-import { UserTooltipContentData } from '../../hooks/useProfileTooltip';
-
-export interface ProfileTooltipContentProps {
-  user: Author;
-  data?: UserTooltipContentData;
-  onMouseEnter?: () => unknown;
-  onMouseLeave?: () => unknown;
-}
-
-export function ProfileTooltipContent({
-  user,
-  data: { rank, tags },
-  onMouseEnter,
-  onMouseLeave,
-}: ProfileTooltipContentProps): ReactElement {
-  return (
-    <div
-      className="flex shrink flex-col font-normal typo-callout"
-      onMouseEnter={onMouseEnter}
-      onMouseLeave={onMouseLeave}
-    >
-      <div className="relative w-fit">
-        <ProfileImageLink user={user} picture={{ size: 'xxlarge' }} />
-        {rank && (
-          <Rank
-            className="absolute -bottom-2 -right-2 h-6 w-6 rounded-8 bg-background-default"
-            rank={rank.currentRank}
-            colorByRank
-            data-testid={rank.currentRank}
-          />
-        )}
-      </div>
-      <ProfileLink
-        className="mt-5 font-bold text-text-primary"
-        href={user.permalink}
-      >
-        {user.name}
-      </ProfileLink>
-      <ProfileLink className="mb-3 text-text-secondary" href={user.permalink}>
-        @{user.username}
-      </ProfileLink>
-      {user.bio && (
-        <p className="mb-3 line-clamp-3 break-words text-text-tertiary">
-          {user.bio}
-        </p>
-      )}
-      {tags?.length ? (
-        <span className="mb-2 text-text-quaternary typo-subhead">
-          Loves reading about
-        </span>
-      ) : null}
-      <TagLinks
-        className={{ container: 'mb-0' }}
-        tags={tags?.map(({ value }) => value) || []}
-      />
-    </div>
-  );
-}
